### 2019-12-12 #1952 @whedon
[PRE REVIEW]: The MFrontGenericInterfaceSupport project
**Submitting author:** @thelfer (<a href="http://orcid.org/0000-0003-2460-5816">Thomas Helfer</a>)
**Repository:** <a href="https://github.com/thelfer/MFrontGenericInterfaceSupport" target ="_blank">https://github.com/thelfer/MFrontGenericInterfaceSupport</a>
**Version:** v1.0.1
