### 2019-12-12 #1952 @kyleniemeyer
@openjournals/dev initially the repo link above was to the specific paper location rather than the repo itself, could that be causing these errors?
