### 2019-12-12 #1952 @kyleniemeyer
Hi @thelfer, thanks for your interest in JOSS. Before we get moving on this, could you clarify how this submission is research software? It looks like it cannot be used by itself, but only by MFront, which is the actual "software" that others would use.
