### 2019-12-12 #1952 @thelfer
> Hi @thelfer, thanks for your interest in JOSS. Before we get moving on this, could you clarify how this submission is research software? It looks like it cannot be used by itself, but only by MFront, which is the actual "software" that others would use.

Maybe the paper is not clear enough on this.

`MFront` generates shared library that must be called by the thermo-mechanical solvers. It is a mature project with has been described and used in many scientific papers.

The `MFrontGenericInterfaceSupportè (`MGIS`) provides tools to the solvers'developers to load those shared libraries and handle the constitutive laws generated by `MFront`. It is a relatively young project and the goal of the paper is to present it.

One may say that:

- `MFront` is meant to be used by the user of thermo-mechanical solvers.
- `MGIS` is meant to be used by the thermo-mechanical solvers' developpers.

In a sense, those projects don't have the same "audience".
