### 2019-12-12 #1952 @thelfer
> @thelfer if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission.

I think I selected

Kevin-Mattheus-Moerman

when submitting the paper.
