### 2019-12-12 #1952 @thelfer
@kyleniemeyer  BTW, when I tried to check the generation of the paper, I got a page telling me that the build was successful but couln't dowload the pdf (file not file).
