### 2019-12-12 #1952 @kyleniemeyer
@thelfer regarding the image errors in the paper:
- we don't support SVG images, so please convert that one to a PDF (or something)
- I'm not sure what's causing the other error, with `MCAS_disc_hole_cyclic_show_axes.png`. That may not actually be an error, but I can't tell.

Regarding the other question, we do some initial checks before assigning an editor to handle the submission—thanks for your patience.
