### 2019-12-12 6df1000
try to fig image handling in whedon
