### 2019-12-12 #1952 @thelfer
> @thelfer regarding the image errors in the paper:
>
> * we don't support SVG images, so please convert that one to a PDF (or something)

@kyleniemeyer This is fixed now.

> * I'm not sure what's causing the other error, with `MCAS_disc_hole_cyclic_show_axes.png`. That may not actually be an error, but I can't tell.

@kyleniemeyer This is quite strange. I just put the whole figure declaration in one line and now it workds. This was not an issue with github engine nor with my version of pandoc.

Anyway, the pdf looks good now. Thank you for your help.
