### 2019-12-13 #1952 @thelfer
> @thelfer - do you think you could reduce your bibtex file to only include the entries that you're actually citing in this paper?

@arfon done.
