### 2019-12-13 #1952 @thelfer
> > @openjournals/dev initially the repo link above was to the specific paper location rather than the repo itself, could that be causing these errors?
>
> Yes, Whedon doesn't know how to handle that.

@arfon Sorry if I did not do things properly. Is there something I could do ?
