### 2019-12-13 #1952 @arfon
>  Is there something I could do ?

No problem. I've tried to improve the JOSS UI to make it clear we want a URL like this:
https://github.com/thelfer/MFrontGenericInterfaceSupport rather than something like this: https://github.com/thelfer/MFrontGenericInterfaceSupport/blob/master/docs/papers/joss/paper.md
