### 2019-12-13 #1952 @thelfer
> No problem. I've tried to improve the JOSS UI to make it clear we want a URL like this:
> https://github.com/thelfer/MFrontGenericInterfaceSupport rather than something like this: https://github.com/thelfer/MFrontGenericInterfaceSupport/blob/master/docs/papers/joss/paper.md

Wouldn't that trigger another issue. I mean, wouldn't it be normal to have several publications (not necessary in joss) in the same repository ?
