### 2019-12-13 #1952 @arfon
> Wouldn't that trigger another issue. I mean, wouldn't it be normal to have several publications (not necessary in joss) in the same repository ?

Possibly, although Whedon only looks for a file called `paper.md`. This hasn't caused any issues in the last ~900 submissions so I think this is an edge case not worth worry about (too much) :-)
