### 2019-12-13 #1952 @Kevin-Mattheus-Moerman
@kyleniemeyer sorry I would be interested but I have a conflict of interest here since one of the authors is a collaborator and friend of mine (Thomas Nagel)
