### 2019-12-15 #1952 @arfon
@thelfer - could you please take a look at [this list of reviewers](https://bit.ly/joss-reviewers) and suggest a few names for people who might be able to review this submission? Please list them without their `@` to avoid pinging them (e.g. `arfon` vs `@arfon`).

Also, @Kevin-Mattheus-Moerman - if you have any suggestions for potential reviewers I'd welcome them 😄
