### 2019-12-16 #1952 @thelfer
@arfon The following people have indicated "mechanical engineering" in their profile. I more or less ordered them by relevance (this is *very* subjective):

- ctdegroot
- shibabrat
- nnadeau
- trallard
- srmnitc
- moorepants
- kyleniemeyer
- CELavecchia
