### 2019-12-17 50fefea
Prepare the release of Version 1.1
