### 2019-12-18 #1952 @arfon
:wave: @CELavecchia & @srmnitc - would you be willing to review this submission for JOSS?
