### 2019-12-23 #1952 @arfon
> @arfon @thelfer I can review this submission. But I am on holidays until the 7th of January. Is it okay if I start after the 7th? If that is the case, I can take this submission.

Yes, that sounds great, thanks! I'll add you as a reviewer now but won't start the actual review thread until I've found a second reviewer.
