### 2019-12-26 #1952 @arfon
@chrisrichardson, @IgorBaratta - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
