### 2020-01-06 #1952 @arfon
:wave: @sebastkm @lindsayad - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
