### 2020-01-07 #1952 @IgorBaratta
> @chrisrichardson, @IgorBaratta - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @arfon , I am back from holidays and now I can review this submission.
I don't have a strong background in mechanics, but I make myself available for the review.
