### 2020-01-07 #1952 @arfon
> Hi @arfon , I am back from holidays and now I can review this submission.
I don't have a strong background in mechanics, but I make myself available for the review.

Great, thanks @IgorBaratta! I'll add you now and start the review.
