### 2020-01-14 3edb4c8
Update documentation. Add the  flag for cmake for   packaging
