+++ docs/web/index.md
@@ -59,6 +59,8 @@ The following bindings are under consideration:
 
 - the `master` branch follows the evolution of the `master` branch of
   the `TFEL` project
+- the `rliv-1.1` follows the evolution of the 3.3.x series of the `TFEL`
+  project.
 - the `rliv-1.0` follows the evolution of the 3.2.x series of the `TFEL`
   project. Note that this branch is **not** compatible with
   `TFEL-3.2.0`.
