+++ docs/web/bibliography.bib
@@ -1,3 +1,31 @@
+@inproceedings{frondelius_mfrontinterface_2019,
+    title={MFrontInterface.jl: MFront material models in Julia{FEM}},
+    author={Tero Frondelius and Thomas Helfer and Ivan Yashchuk and Joona Vaara  and Anssi Laukkanen},
+    editor={H. Koivurova and A. H. Niemi},
+    booktitle={Proceedings of the 32nd Nordic Seminar on Computational Mechanics},
+    year={2019},
+    place={Oulu}
+}
+
+@misc{bleyer_fenics_2019,
+	title = {{FEniCS} and {MFront} for complex non linear solid mechanics simulation},
+	url = {https://www.researchgate.net/publication/333089270_FEniCS_and_MFront_for_complex_non_linear_solid_mechanics_simulation},
+	abstract = {{ResearchGate} is a network dedicated to science and research. Connect, collaborate and discover scientific publications, jobs and conferences. All for free.},
+	author = {Bleyer, Jérémy and Helfer, Thomas},
+	date = {2019},
+	langid = {english},
+	doi = {10.13140/RG.2.2.35501.54247},
+	file = {Snapshot:/home/th202608/.zotero/zotero/vzpzva46.default/zotero/storage/LHXG9FUF/333089270_FEniCS_and_MFront_for_complex_non_linear_solid_mechanics_simulation.html:text/html}
+}
+
+@online{bleyer_elasto-plastic_2019,
+	title = {Elasto-plastic analysis implemented using the {MFront} code generator},
+	url = {https://comet-fenics.readthedocs.io/en/latest/demo/plasticity_mfront/plasticity_mfront.py.html},
+	titleaddon = {Numerical tours of continuum mechanics using {FEniCS}},
+	author = {Bleyer, Jérémy and Helfer, Thomas},
+	date = {2019}
+}
+
 @article{Lester2017,
 title = {Verification of the Skorohod-Olevsky Viscous Sintering (SOVS) Model},
 author = {Lester, Brian T.},
