### 2020-01-14 b802729
Add the `USE_EXTERNAL_COMPILER_FLAGS` flag for cmake for  `conda` packaging
