+++ cmake/modules/compiler.cmake
@@ -156,25 +156,37 @@ endif(CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
 set(CMAKE_C_FLAGS   "${COMPILER_FLAGS} ${COMPILER_CFLAGS} ${COMPILER_C_WARNINGS}")
 set(CMAKE_CXX_FLAGS "${VISIBILITY_FLAGS} ${COMPILER_WARNINGS} ${COMPILER_FLAGS} ${COMPILER_CXXFLAGS}")
 
-if(CMAKE_BUILD_TYPE STREQUAL "Profiling")
-  set(CMAKE_C_FLAGS_PROFILING "${OPTIMISATION_FLAGS} ${CMAKE_C_FLAGS_PROFILING}")
-  set(CMAKE_CXX_FLAGS_PROFILING "${OPTIMISATION_FLAGS} ${CMAKE_CXX_FLAGS_PROFILING}")
-  if(NOT enable-portable-build)
-    set(CMAKE_C_FLAGS_PROFILING "${OPTIMISATION_FLAGS_MARCH} ${CMAKE_C_FLAGS_PROFILING}")
-    set(CMAKE_CXX_FLAGS_PROFILING "${OPTIMISATION_FLAGS_MARCH} ${CMAKE_CXX_FLAGS_PROFILING}")
-  endif(NOT enable-portable-build)
-elseif(CMAKE_BUILD_TYPE STREQUAL "Release")
-  set(CMAKE_C_FLAGS_RELEASE     "${OPTIMISATION_FLAGS} ${CMAKE_C_FLAGS_RELEASE}")
-  set(CMAKE_CXX_FLAGS_RELEASE   "${OPTIMISATION_FLAGS} ${CMAKE_CXX_FLAGS_RELEASE}")
-  if(NOT enable-portable-build)
-    set(CMAKE_C_FLAGS_RELEASE "${OPTIMISATION_FLAGS_MARCH} ${CMAKE_C_FLAGS_RELEASE}")
-    set(CMAKE_CXX_FLAGS_RELEASE "${OPTIMISATION_FLAGS_MARCH} ${CMAKE_CXX_FLAGS_RELEASE}")
-  endif(NOT enable-portable-build)
-else(CMAKE_BUILD_TYPE STREQUAL "Profiling")
-  set(CMAKE_C_FLAGS             "${OPTIMISATION_FLAGS} ${CMAKE_C_FLAGS}")
-  set(CMAKE_CXX_FLAGS           "${OPTIMISATION_FLAGS} ${CMAKE_CXX_FLAGS}")
-  if(NOT enable-portable-build)
-    set(CMAKE_C_FLAGS "${OPTIMISATION_FLAGS_MARCH}   ${CMAKE_C_FLAGS}")
-    set(CMAKE_CXX_FLAGS "${OPTIMISATION_FLAGS_MARCH} ${CMAKE_CXX_FLAGS}")
-  endif(NOT enable-portable-build)
-endif(CMAKE_BUILD_TYPE STREQUAL "Profiling")
+# This option has been added for building conda package.
+# It circumvents the following issue: `cmake` discards `Boost_INCLUDEDIRS`
+# which is equal to `$PREFIX/include`. The same applies to
+# `PYTHON_INCLUDEDIRS`.
+#
+# `conda` adds `-I$PREFIX/include` in the `CFLAGS` and `CXXFLAGS`,
+# thus, using those variables `CFLAGS` and `CXXFLAGS` solves the issue and is more
+# consistent with other `conda` packages.
+#
[…]
