### 2020-01-14 6ee0a70
Add the `USE_EXTERNAL_COMPILER_FLAGS` flag for cmake for  `conda` packaging
