+++ cmake/modules/compiler.cmake
@@ -167,9 +167,6 @@ else(HAVE_NORETURN_ATTRIBUTE)
   MESSAGE(STATUS "disabling [[ noreturn ]] attribute")
 endif(HAVE_NORETURN_ATTRIBUTE)
 			  
-set(CMAKE_C_FLAGS   "${COMPILER_FLAGS} ${COMPILER_CFLAGS} ${COMPILER_C_WARNINGS}")
-set(CMAKE_CXX_FLAGS "${VISIBILITY_FLAGS} ${COMPILER_WARNINGS} ${COMPILER_FLAGS} ${COMPILER_CXXFLAGS}")
-
 # This option has been added for building conda package.
 # It circumvents the following issue: `cmake` discards `Boost_INCLUDEDIRS`
 # which is equal to `$PREFIX/include`. The same applies to
