### 2020-01-19 thelfer/MFrontGenericInterfaceSupport#17 @srmnitc
Installation instruction and dependencies
This is part of the review of JOSS paper https://github.com/openjournals/joss-reviews/issues/2003

From the JOSS paper and associated documentation [webpage](https://thelfer.github.io/mgis/web/index.html), it is difficult to find any documentation on how to install or compile MGIS. Furthermore, the dependencies are also not documented in an easily accessible way.
