### 2020-01-19 thelfer/MFrontGenericInterfaceSupport#18 @srmnitc
Running tests
This is part of the review of JOSS paper https://github.com/openjournals/joss-reviews/issues/2003

A folder named tests is present in the repo, however, no documentation is provided on how to run the tests
