### 2020-01-19 thelfer/MFrontGenericInterfaceSupport#19 @srmnitc
API Documentation
This is part of the review of JOSS paper https://github.com/openjournals/joss-reviews/issues/2003

The [webpage](https://thelfer.github.io/mgis/web/bindings-cxx.html) mentions that API documentation can be generated using Doxygen. However, when trying to generate the documentation, the following error is generated-

```
warning: tag INPUT: input source `@CMAKE_SOURCE_DIR@' does not exist
Exiting...
```
If any further configuration is required, it needs to be documented.
