### 2020-01-19 thelfer/MFrontGenericInterfaceSupport#20 @srmnitc
Community guidelines absent
This is part of the review of JOSS paper https://github.com/openjournals/joss-reviews/issues/2003

Community guidelines as present in the JOSS paper - `Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support` are absent.
