### 2020-01-19 #2003 @srmnitc
@arfon  @thelfer I have completed a first round of review. Thanks @thelfer for the repository and documentation. The documentation, especially the associated webpage documents various examples and usage quite well. My issue was with initial setup. I have raised four issues at the repository - https://github.com/thelfer/MFrontGenericInterfaceSupport/issues/17, https://github.com/thelfer/MFrontGenericInterfaceSupport/issues/18, https://github.com/thelfer/MFrontGenericInterfaceSupport/issues/19, and https://github.com/thelfer/MFrontGenericInterfaceSupport/issues/20.  For me, it was hard to see how to set up or install the software and use it. Once the issues are fixed, I can move on to check the functionality.

There are also few things for which I have not opened an issue. This has to do with the JOSS paper-

- The JOSS paper has some typos which need to fixed.
- The first page of the paper has numerous short paragraphs, which in my opinion affects the readability.
- Another point is about citations. Statements in the paper such as `This is error-prone and may lead to spurious or even worse inexact results` and `While this causes memory fragmentation,
this is the most frequent choice`, would be reinforced with suitable citations.
- Finally, there are multiple examples provided. This is one for the editor @arfon. These examples could be moved to a documentation in order to keep the JOSS paper short and concise.
