### 2020-01-20 thelfer/MFrontGenericInterfaceSupport#21 @srmnitc
Author affiliations are not complete
This is part of the review of JOSS paper https://github.com/openjournals/joss-reviews/issues/2003

Author affiliations should have complete addresses
