### 2020-01-23 #2003 @arfon
> Finally, there are multiple examples provided. This is one for the editor @arfon. These examples could be moved to a documentation in order to keep the JOSS paper short and concise.

Thanks @srmnitc for your feedback here. To respond directly to ☝️this comment - I think this is probably OK as-is. We generally try to avoid lengthy discussions of e.g. the API structure (something that should be in docs) but some examples such as these are reasonable I think.
