### 2020-01-24 #2003 @IgorBaratta
I started the review by trying the _experimental_ Julia binding, and the installation of the package was straightforward. However, for the C++ library itself, I couldn't find any documentation of the installation process.  So at this point, I could not honestly evaluate the **Functionality**.

Concerning the **Documentation**, some items are missing or not easily accessible:
- Community guidelines
- Automated tests - Are the tests automated or are there any manual steps to run them?
- Installation instructions

Some examples are given in https://thelfer.github.io/mgis/web/bindings-cxx.html , and it seems to be a good start for using the software, but I miss a standalone example I could just run and modify. Or, are there any present in the test folder? Nevertheless, I couldn't find any documented.
