### 2020-01-28 #2003 @thelfer
@IgorBaratta, @srmnitc Sorry for not responding earlier. Thank you for this first round of review and those very constructive feed-backs.

I just uploaded a new version of the `INSTAL-cmake.md` file which now fully describes the installation process. Let me know if this file is self-explanatory (it requires some knowledge of `cmake` tough).
