### 2020-01-28 #2003 @thelfer
> * Automated tests - Are the tests automated or are there any manual steps to run them?

To run the tests, you must type `make check` in the `build` directory. Tests are only available if the `TFEL/MFront` package is available. As explained in the paper, without `TFEL/MFront`, the package is barely useless. Yet, technically, `TFEL/MFront` is *not* a dependency of `MGIS`.
