### 2020-01-28 #2003 @thelfer
> However, for the C++ library itself, I couldn't find any documentation of the installation process.

The `C++` library is *always* built. The `Julia` bindings just wraps the `C++` methods using the `CxxWrap` package.
