### 2020-01-28 #2003 @thelfer
> * Community guidelines

@IgorBaratta  Are those required ?
