### 2020-01-28 #2003 @thelfer
> Some examples are given in https://thelfer.github.io/mgis/web/bindings-cxx.html , and it seems to be a good start for using the software

@IgorBaratta  Indeed, this document is the main entry point for the `C++` library. The `MGIS` library is indeed quite small so it provides a convenient introduction/overview.

Interested users then have a look at the `tests` directory where examples are provided.

 I do appreciate your feed-backs on the documentation.
