### 2020-01-28 thelfer/MFrontGenericInterfaceSupport#19 @thelfer
The `Doxygen` documentation can be generated as part of the build process. You must pass the `-Denable-doxygen-doc=ON` flag to cmake to enable it (See `INSTAL-cmake.md` for details). The documentation is generated by:

```bash
$ make doc-html
```
