### 2020-01-28 thelfer/MFrontGenericInterfaceSupport#21 @thelfer
@srmnitc Could you be more specific ? Is the issue only related to the JOSS paper ? If so, each authors seems to have properly filled its affiliations in the paper, don't they ?
