### 2020-01-28 thelfer/MFrontGenericInterfaceSupport#18 @thelfer
Running the test is part of the build process. The tests can be run by calling `ctest` (as usual in a `cmake` project), or by calling :

~~~~
$ make check
~~~~

at the top of the build directory.

**Note** The tests are only built if `TFEL/MFront` has been detected by `cmake`.
