### 2020-01-28 thelfer/MFrontGenericInterfaceSupport#17 @thelfer
@srmnitc Thanks for this feed-back. I updated the `INSTALL-cmake.md` file to give an overview of the installation process. A new page has been added to the website: https://thelfer.github.io/mgis/web/install.html
