### 2020-01-28 thelfer/MFrontGenericInterfaceSupport#21 @srmnitc
@thelfer Yes this is only related to the paper.  No, they are incomplete. For example, number 3 Wärtsilä, Number 4, Number 6 etc. The complete address of affiliated institution should be present. All of the already published JOSS papers seems to follow this.
