### 2020-01-29 8c3c043
Updated affiliations
Fixes #21
