### 2020-01-29 thelfer/MFrontGenericInterfaceSupport#21 @thelfer
@srmnitc Can I close the issue ?
