### 2020-01-29 thelfer/MFrontGenericInterfaceSupport#21 @srmnitc
1, 7 and 8 still not fixed
