### 2020-01-29 3ff720f
Paper update. Installation guide update
