+++ INSTALL-cmake.md
@@ -5,8 +5,8 @@ These are generic installation instructions.
 
 cmake attempts to guess correct values for various system-dependent
 variables used during compilation.  It uses those values to create 
-a `Makefile' in each directory of the package.  It may also create one
-or more `.h' files containing system-dependent definitions..
+a `Makefile` in each directory of the package.  It may also create one
+or more `.h` files containing system-dependent definitions..
 
 The simplest way to compile this package is:
 
@@ -25,16 +25,16 @@ The simplest way to compile this package is:
    Running `cmake` takes awhile.  While running, it prints some
    messages telling which features it is checking for.
 
-2. Type `make' to compile the package.
+2. Type `make` to compile the package.
 
-3. Optionally, type `make check' to run any self-tests that come with
+3. Optionally, type `make check` to run any self-tests that come with
   the package.
 
-4. Type `make install' to install the programs and any data files and
+4. Type `make install` to install the programs and any data files and
   documentation.
 
 5. You can remove the program binaries and object files from the source
-  code directory by typing `make clean'.
+  code directory by typing `make clean`.
 
 Options
 =======
@@ -56,10 +56,10 @@ Options
 - `enable-doxygen-doc`: enable the generation of the API documentation
   using with `Doxygen`.
 
-cmake usefull variables
+`cmake` usefull variables
 =======================
 
-- `CMAKE_BUILD_TYPE`           : two values are supported 'Release' and 'Debug'
+- `CMAKE_BUILD_TYPE`           : two values are supported `Release` and `Debug`
 - `CASTEM_INSTALL_PATH`        : specify where the castem has been installed
 - `CMAKE_TOOLCHAIN_FILE`       : specify a tool chain file
 - `Python_ADDITIONAL_VERSIONS` : select the `python` version to
@@ -68,8 +68,8 @@ cmake usefull variables
   			         the revision version or the detection
   			         fails.
 
-cmake typical usage
-===================
+`cmake` typical usage
+=====================
 
 mkdir build
 cd build
