### 2020-01-29 thelfer/MFrontGenericInterfaceSupport#21 @thelfer
Updated my own affiliation. Waiting for Thomas and Dmitri updates.
