### 2020-01-29 7141486
Installation guide update. Add a Contributing section on the main page. Add the CONTRIBUTING.md file
