### 2020-01-29 357d2c1
Update the contributing guide. Fix Issue #20
