### 2020-01-29 thelfer/MFrontGenericInterfaceSupport#20 @thelfer
@srmnitc I adapted the `CONTRIBUTE.md` file for the `html5boilerplate` project which I found very clear. I added this new `CONTRIBUTE.md` in the root directory of the project and updated the website. Need to do the same for all my other projects.
