### 2020-01-30 thelfer/MFrontGenericInterfaceSupport#21 @thelfer
@srmnitc Affliations shall be correct by now.
