### 2020-02-24 2b3dbb1
update affiliation according to CEA new rules
