### 2020-02-27 #2003 @thelfer
@arfon @IgorBaratta T. Nagel, one of the co-author, just propose to update the `OpenGeoSys` part with a more meaningful example. Would you mind if we do this change ? (https://github.com/thelfer/MFrontGenericInterfaceSupport/pull/26)
