+++ cmake/modules/compiler.cmake
@@ -184,3 +184,4 @@ else(USE_EXTERNAL_COMPILER_FLAGS)
     endif(NOT enable-portable-build)
   endif(CMAKE_BUILD_TYPE STREQUAL "Profiling")
 endif(USE_EXTERNAL_COMPILER_FLAGS)
+
