### 2020-02-27 #2003 @IgorBaratta
> > * Community guidelines
>
> @IgorBaratta Are those required ?

I think so, it's on the review checklist, and it's one of the boxes I still have to check.
