+++ bindings/c/include/MGIS/Behaviour/Variable.h
@@ -15,6 +15,10 @@
 #ifndef LIB_MGIS_BEHAVIOUR_VARIABLE_H
 #define LIB_MGIS_BEHAVIOUR_VARIABLE_H
 
+#ifdef __cplusplus
+extern "C" {
+#endif /*  __cplusplus */
+
 /*!
  * \brief type of a variable
  */
@@ -24,5 +28,19 @@ typedef enum {
   MGIS_BV_STENSOR = 2,
   MGIS_BV_TENSOR  = 3
 } mgis_bv_VariableType;
+
+/*!
+ * \return the size of an array able to store the variable
+ * \param [out] s: size
+ * \param [in] h: modelling hypothesis
+ * \param [in] t: variable type
+ */
+MGIS_C_EXPORT mgis_status mgis_bv_get_variable_size(mgis_size_type* const,
+						    const char* const,
+						    const mgis_bv_VariableType);
+
+#ifdef __cplusplus
+}  // end of extern "C"
+#endif
 
 #endif /* LIB_MGIS_BEHAVIOUR_VARIABLE_H */
