### 2020-03-02 c4c2ae2
Fix Issue #25. Documentation update.
