+++ cmake/modules/mgis.cmake
@@ -1,19 +1,27 @@
-macro(mgis_header dir file)
+function(mgis_header dir file)
   install(FILES ${dir}/${file}
     DESTINATION "include/${dir}")
-endmacro(mgis_header)
+endfunction(mgis_header)
 
-macro(mgis_library name)
+function(mgis_library name)
   if(${ARGC} LESS 2)
     message(FATAL_ERROR "mgis_library_internal : no source specified")
   endif(${ARGC} LESS 2)
   add_library(${name} SHARED ${ARGN})
   if(WIN32)
-    install(TARGETS ${name} DESTINATION bin)
+    install(TARGETS ${name} EXPORT ${name}
+            DESTINATION bin)
   else(WIN32)
-    install(TARGETS ${name}
-      DESTINATION lib${LIB_SUFFIX})
+    install(TARGETS ${name} EXPORT ${name}
+            DESTINATION lib${LIB_SUFFIX})
   endif(WIN32)
+  if(MGIS_APPEND_SUFFIX)
+    set(export_install_path "share/mgis-${MGIS_SUFFIX}/cmake")
+  else(MGIS_APPEND_SUFFIX)
+    set(export_install_path "share/mgis/cmake")
+  endif(MGIS_APPEND_SUFFIX)
+  install(EXPORT ${name} DESTINATION ${export_install_path}
+          NAMESPACE mgis:: FILE ${name}Config.cmake)
   if(enable-static)
     add_library(${name}-static STATIC ${ARGN})
     set_target_properties(${name}-static PROPERTIES OUTPUT_NAME "${name}")
@@ -35,4 +43,4 @@ macro(mgis_library name)
       install(TARGETS ${name}-static DESTINATION lib${LIB_SUFFIX})
     endif(WIN32)
   endif(enable-static)
-endmacro(mgis_library)
+endfunction(mgis_library)
