+++ .github/workflows/ccpp.yml
@@ -5,12 +5,12 @@ on: [push]
 jobs:
   build:
 
-    runs-on: [ubuntu-latest, windows-latest, macOS-latest]
+    runs-on: ubuntu-latest
     
     steps:
     - uses: actions/checkout@v2
     - name: cmake
-      run: cmake -DCMAKE_BUILD_TYPE=Release
+      run: cmake -DCMAKE_BUILD_TYPE=Release .
     - name: make
       run: make
     - name: make check
