+++ .github/workflows/ccpp.yml
@@ -5,7 +5,20 @@ on: [push]
 jobs:
   build:
 
-    runs-on: [ubuntu-latest, windows-latest, macOS-latest]
+    runs-on: ubuntu-latest
+    
+    steps:
+    - uses: actions/checkout@v2
+    - name: cmake
+      run: cmake -DCMAKE_BUILD_TYPE=Release .
+    - name: make
+      run: make
+    - name: make check
+      run: make check
+
+  build:
+
+    runs-on: macOS-latest
     
     steps:
     - uses: actions/checkout@v2
