+++ .github/workflows/ccpp.yml
@@ -10,20 +10,7 @@ jobs:
     steps:
     - uses: actions/checkout@v2
     - name: cmake
-      run: cmake -DCMAKE_BUILD_TYPE=Release .
-    - name: make
-      run: make
-    - name: make check
-      run: make check
-
-  build:
-
-    runs-on: macOS-latest
-    
-    steps:
-    - uses: actions/checkout@v2
-    - name: cmake
-      run: cmake -DCMAKE_BUILD_TYPE=Release .
+      run: cmake -DCMAKE_BUILD_TYPE=Release . -Denable-c-bindings=ON  -Denable-fortran-bindings=ON -Denable-python-bindings=OFF -Denable-portable-build=ON -Denable-julia-bindings=OFF
     - name: make
       run: make
     - name: make check
