### 2020-03-13 cdab675
Finalize work on Issue #33: update c and fortran bindings, add unit tests, update release notes
