### 2020-03-25 thelfer/MFrontGenericInterfaceSupport#35 @srmnitc
Issues with paper
This is part of the review of JOSS paper openjournals/joss-reviews#2003

- The JOSS paper has some typos which need to fixed.
- The first page of the paper has numerous short paragraphs, which in my opinion affects the readability.
- Another point is about citations. Statements in the paper such as `This is error-prone and may lead to spurious or even worse inexact results` and `While this causes memory fragmentation, this is the most frequent choice, would be reinforced with suitable citations`.
