### 2020-03-25 thelfer/MFrontGenericInterfaceSupport#35 @thelfer
Hi @srmnitc,

> Statements in the paper such as This is error-prone and may lead to spurious or even
> worse inexact results and While this causes memory fragmentation, this is the most frequent
> choice, would be reinforced with suitable citations.

This is indeed a difficult point coming most notably from our own experiences of building complex simulations. I don't know if I can find any citations on that matter.

Would it be ok if we introduce proper word of cautions, such as: "According to the authors experience...", "One drawback of this approach is that it may lead to memory fragmentation"
