### 2020-03-25 1f88f30
Further additions and minor corrections to remaining sections.
