### 2020-03-26 thelfer/MFrontGenericInterfaceSupport#35 @srmnitc
I think you could also specify some reason as to why it is error prone and so on.
