### 2020-03-31 ae53547
Work on Issue #35. Code clean-up
