### 2020-03-31 thelfer/MFrontGenericInterfaceSupport#35 @thelfer
For the third point, here is a small paragraph which highlighs how traditional approaches are error-prone:

> In the authors' experience, this is error-prone in particular for
> inexperienced users and may lead to spurious or even worse inexact
> results. For example, when material properties such as the Young Modulus
> or the Poission ratio, are required by the behaviour, those are
> generally defined in the solver input file by an array of values and
> potentiel checks are limited to the size of the array. An user may thus
> invert two material properties. If those two material properties have
> the same order of magnitudes, computations might lead to a physically
> consistent result, altough a wrong result.
