### 2020-03-31 4efc2a3
Few fixes in the paper. Rescale the JuliaFEM figure
