### 2020-03-31 cd034b5
Remove remark about memory fragmentation
