### 2020-03-31 thelfer/MFrontGenericInterfaceSupport#35 @thelfer
@srmnitc

> * The JOSS paper has some typos which need to fixed.

I tried to fix that with the help of the co-authors and a colleague who is a native english speaker. English is hopefuly acceptable now.

> * The first page of the paper has numerous short paragraphs, which in my opinion affects the readability.

Apart from the first line, the other paragraphs have more than 7 lines. Each develops a specific idea, which is a common convention in French.

> * Another point is about citations. Statements in the paper such as `This is error-prone and may lead to spurious or even worse inexact results` and `While this causes memory fragmentation, this is the most frequent choice, would be reinforced with suitable citations`.

The statement "This is error-prone" has been discussed more in details.
The statement about memory fragmentation has been removed because detailled discussion of this point is out of the scope of the paper.
