### 2020-04-03 thelfer/MFrontGenericInterfaceSupport#35 @srmnitc
Thanks @thelfer , everything looks okay now. One last point - maybe Mohr-Coulumb yield criterion needs a citation?
