### 2020-04-03 97160ad
Add references regarding the Mohr-Coulomb criterion
