+++ docs/papers/joss/bibliography.bib
@@ -1,3 +1,30 @@
+@article{abbo_smooth_1995,
+	title = {A smooth hyperbolic approximation to the Mohr-Coulomb yield criterion},
+	volume = {54},
+	issn = {0045-7949},
+	url = {http://www.sciencedirect.com/science/article/pii/0045794994003395},
+	doi = {10.1016/0045-7949(94)00339-5},
+	abstract = {The Mohr-Coulomb yield criterion is used widely in elastoplastic geotechnical analysis. There are computational difficulties with this model, however, due to the gradient discontinuities which occur at both the edges and the tip of the hexagonal yield surface pyramid. It is well known that these singularities often cause stress integration schemes to perform inefficiently or fail. This paper describes a simple hyperbolic yield surface that eliminates the singular tip from the Mohr-Coulomb surface. The hyperbolic surface can be generalized to a family of Mohr-Coulomb yield criteria which are also rounded in the octahedral plane, thus eliminating the singularities that occur at the edge intersections as well. This type of yield surface is both continuous and differentiable at all stress states, and can be made to approximate the Mohr-Coulomb yield function as closely as required by adjusting two parameters. The yield surface and its gradients are presented in a form which is suitable for finite element programming with either explicit or implicit stress integration schemes. Two efficient {FORTRAN} 77 subroutines are given to illustrate how the new yield surface can be implemented in practice.},
+	pages = {427--441},
+	number = {3},
+	journaltitle = {Computers \& Structures},
+	shortjournal = {Computers \& Structures},
+	author = {Abbo, A. J. and Sloan, S. W.},
+	date = {1995-02-03},
+	file = {Abbo et Sloan - 1995 - A smooth hyperbolic approximation to the Mohr-Coul.pdf:/home/th202608/.zotero/zotero/vzpzva46.default/zotero/storage/HW8PAKZN/Abbo et Sloan - 1995 - A smooth hyperbolic approximation to the Mohr-Coul.pdf:application/pdf;ScienceDirect Snapshot:/home/th202608/.zotero/zotero/vzpzva46.default/zotero/storage/MV5E75AK/0045794994003395.html:text/html}
+}
+
+@InProceedings{zienkiewicz-pande-77,
+  Author = {Zienkiewicz, O.C. and Pande, G.N.},
+  Title = {{Some Useful Forms of Isotropic Yield Surfaces for Soil and Rock Mechanics}},
+  BookTitle = {Finite Elements in Geomechanics},
+  editor = {Gudehus, G.},
+  Pages = {179--190},
+  Publisher = {John Wiley & Sons, Inc, New York, London, Sydney},
+  Year = {1977}
+}
+
+
 @inproceedings{frondelius_mfrontinterface_2019,
     title={MFrontInterface.jl: MFront material models in Julia{FEM}},
     author={Tero Frondelius and Thomas Helfer and Ivan Yashchuk and Joona Vaara  and Anssi Laukkanen},
