### 2020-04-03 c6f66e9
Remove useless parentheses
