### 2020-04-03 thelfer/MFrontGenericInterfaceSupport#35 @thelfer
Done ! Thanks again for the review.
