### 2020-04-03 #2003 @thelfer
Hi @arfon, @srmnitc @IgorBaratta
I hope everyone is fine.
I just closed this issue:  https://github.com/thelfer/MFrontGenericInterfaceSupport/issues/35
Here there anything I can do to help the review process ?
Regards,
Thomas
