### 2020-04-03 #2003 @srmnitc
With the last review, I am happy with the overall status of the paper. The main part where I struggled was instructions to set up and use the code, but now the documentation is much better. Also, for the specialist audience, it might be easier. I still think the paper could be improved a bit more in terms of sentence construction and language, but it is in overall good shape. With that said, I have completed my review and recommend the paper for publication @arfon . Thanks @thelfer for the modifications and feedback. I will still post some issues for the general improvement of the code, but they do not block the publication of the paper. Thanks and hope everybody is healthy and in good spirits!
