### 2020-04-08 #2003 @thelfer
Just fixed a typo in the acknowledgements.
