### 2020-04-08 f1955a8
Prepare to integrate Jérémy Bleyer work
