+++ bindings/python/src/CMakeLists.txt
@@ -1,22 +1,3 @@
-# store the location of the python modules in the MGIS_PYTHON_MODULES_INSTALL_DIRECTORY variable
-if(MGIS_APPEND_SUFFIX)
-  if(WIN32)
-    set(MGIS_PYTHON_MODULES_INSTALL_DIRECTORY
-        bin/python${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}/site-packages/mgis_${MGIS_SUFFIX_FOR_PYTHON_MODULES})
-  else(WIN32)
-    set(MGIS_PYTHON_MODULES_INSTALL_DIRECTORY
-        "lib${LIB_SUFFIX}/python${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}/site-packages/mgis_${MGIS_SUFFIX_FOR_PYTHON_MODULES}")
-  endif(WIN32)
-else(MGIS_APPEND_SUFFIX)
-  if(WIN32)
-    set(MGIS_PYTHON_MODULES_INSTALL_DIRECTORY
-        "bin/python${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}/site-packages/mgis")
-  else(WIN32)
-    set(MGIS_PYTHON_MODULES_INSTALL_DIRECTORY
-        "lib${LIB_SUFFIX}/python${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}/site-packages/mgis")
-  endif(WIN32)
-endif(MGIS_APPEND_SUFFIX)
-
 function(mgis_python_module name output_name)
   if(${ARGC} LESS 3)
     message(FATAL_ERROR "mgis_python_module: no source specified")
@@ -78,8 +59,3 @@ mgis_python_module(mgis_behaviour behaviour
  MaterialStateManager.cxx
  Integrate.cxx
  FiniteStrainSupport.cxx)
-
-install(FILES __init__.py
-        DESTINATION ${MGIS_PYTHON_MODULES_INSTALL_DIRECTORY}
-        COMPONENT python_bindings)
-
