### 2020-04-09 612b4d8
Add support for Jupyter notebooks for generating the website
