+++ cmake/modules/jupyter-nbconvert.cmake
@@ -1,0 +1,15 @@
+find_program(JUPYTER_NBCONVERT NAMES jupyter-nbconvert jupyter-nbconvert.exe)
+mark_as_advanced(JUPYTER_NBCONVERT)
+if(JUPYTER_NBCONVERT)
+  message(STATUS "Enabling generation of documentation based on jupyter-nbconvert")
+  set(MGIS_HAVE_JUPYTER_NBCONVERT ON)
+elseif(JUPYTER_NBCONVERT)
+  message(STATUS "Disabling generation of documentation based on jupyter-nbconvert")
+  set(MGIS_HAVE_JUPYTER_NBCONVERT OFF)
+endif(JUPYTER_NBCONVERT)
+
+if(JUPYTER_NBCONVERT)
+  message(STATUS "jupyter-nbconvert:          ${JUPYTER_NBCONVERT}")
+elseif(JUPYTER_NBCONVERT)
+  message(STATUS "jupyter-nbconvert:          not found")
+endif(JUPYTER_NBCONVERT)
