+++ CMakeLists.txt
@@ -231,14 +231,22 @@ endif(enable-doxygen-doc)
 
 option(enable-website "enable generation of the website" ON)
 if(enable-website)
+  set(MGIS_GENERATE_WEBSITE OFF)
   # Looking for pandoc (http://johnmacfarlane.net/pandoc)
   include(cmake/modules/pandoc.cmake)
-  if(HAVE_PANDOC)
+  if(MGIS_HAVE_PANDOC)
+    set(MGIS_GENERATE_WEBSITE ON)
+  endif(MGIS_HAVE_PANDOC)
+  # Looking for jupyter-nbconvert
+  include(cmake/modules/jupyter-nbconvert.cmake)
+  if(MGIS_HAVE_JUPYTER_NBCONVERT)
+    set(MGIS_GENERATE_WEBSITE ON)
+  endif(MGIS_HAVE_JUPYTER_NBCONVERT)
+  if(MGIS_GENERATE_WEBSITE)
     list(APPEND CPACK_COMPONENTS_ALL website)
     set(CPACK_COMPONENT_WEBSITE_DESCRIPTION
       "Contains the TFEL web site")
-    set(MGIS_GENERATE_WEBSITE ON)
-  endif(HAVE_PANDOC)
+  endif(MGIS_GENERATE_WEBSITE)
 else(enable-website)
   message(STATUS "Generation of the TFEL website disabled")
 endif(enable-website)
