+++ cmake/modules/pandoc.cmake
@@ -6,10 +6,10 @@ mark_as_advanced(PANDOC_CITEPROC)
 mark_as_advanced(PANDOC_CROSSREF)
 if(PANDOC AND PANDOC_CITEPROC AND PANDOC_CROSSREF)
   message(STATUS "Enabling generation of documentation based on pandoc")
-  set(HAVE_PANDOC ON)
+  set(MGIS_HAVE_PANDOC ON)
 elseif(PANDOC AND PANDOC_CITEPROC AND PANDOC_CROSSREF)
   message(STATUS "Disabling generation of documentation based on pandoc")
-  set(HAVE_PANDOC OFF)
+  set(MGIS_HAVE_PANDOC OFF)
 endif(PANDOC AND PANDOC_CITEPROC AND PANDOC_CROSSREF)
 
 if(PANDOC)
