### 2020-04-09 d318f3e
Fixed typo in CMakeLists for fenics
