+++ bindings/python/mgis/fenics/CMakeLists.txt
@@ -1,6 +1,6 @@
 function(mgis_fenics_python_script file)
   install(FILES ${file}
-          DESTINATION ${MGIS_PYTHON_MODULES_INSTALL_DIRECTORY}/fencis
+          DESTINATION ${MGIS_PYTHON_MODULES_INSTALL_DIRECTORY}/fenics
           COMPONENT python_bindings)
 endfunction(mgis_fenics_python_script)
 
