+++ bindings/python/mgis/fenics/nonlinear_problem.py
@@ -1,0 +1,529 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+MFrontNonlinearProblem and MFrontOptimisation classes
+
+@author: Jeremy Bleyer, Ecole des Ponts ParisTech,
+Laboratoire Navier (ENPC,IFSTTAR,CNRS UMR 8205)
+@email: jeremy.bleyer@enpc.f
+"""
+from dolfin import *
+from .utils import *
+from .gradient_flux import *
+import mgis.behaviour as mgis_bv
+from ufl import shape
+
+SymGrad = {mgis_bv.Hypothesis.Tridimensional: lambda u: sym(grad(u)),
+           mgis_bv.Hypothesis.PlaneStrain: lambda u: sym(grad(u)),
+           mgis_bv.Hypothesis.Axisymmetrical: lambda r, u: sym(axi_grad(r, u))}
+Grad =  {mgis_bv.Hypothesis.Tridimensional: lambda u: grad(u),
+         mgis_bv.Hypothesis.PlaneStrain: lambda u: grad(u),
+         mgis_bv.Hypothesis.Axisymmetrical: lambda r, u: axi_grad(r, u)}
+Id_Grad = {mgis_bv.Hypothesis.Tridimensional: lambda u: Identity(3)+grad(u),
+           mgis_bv.Hypothesis.PlaneStrain: lambda u: Identity(2)+grad(u),
+           mgis_bv.Hypothesis.Axisymmetrical: lambda r, u: Identity(3)+axi_grad(r, u)}
+
+predefined_gradients = {
+    "Strain": SymGrad,
+    "TemperatureGradient": Grad,
+    "DeformationGradient": Id_Grad}
+
+predefined_external_state_variables = {"Temperature": Constant(293.15)}
+
+def list_predefined_gradients():
+    print("'TFEL gradient name'   (Available hypotheses)")
+    print("---------------------------------------------")
+    for (g, value) in predefined_gradients.items():
+        print("{:22} {}".format("'{}'".format(g), tuple(str(v) for v in value.keys())))
+    print("")
+
+class AbstractNonlinearProblem:
+    """
+    This class handles the definition and resolution of a nonlinear problem associated with a `MFrontNonlinearMaterial`.
+    """
+    def __init__(self, u, material, quadrature_degree=2, bcs=None):
+        self.u = u
+        self.V = self.u.function_space()
+        self.u_ = TestFunction(self.V)
+        self.du = TrialFunction(self.V)
+        self.mesh = self.V.mesh()
+        self.material = material
+        self.axisymmetric = self.material.hypothesis==mgis_bv.Hypothesis.Axisymmetrical
+        self.integration_type = mgis_bv.IntegrationType.IntegrationWithConsistentTangentOperator
+
+        self.quadrature_degree = quadrature_degree
+
+
+        cell = self.mesh.ufl_cell()
+        W0e = get_quadrature_element(cell, self.quadrature_degree)
+        # scalar quadrature space
+        self.W0 = FunctionSpace(self.mesh, W0e)
+        # compute Gauss points numbers
+        self.ngauss = Function(self.W0).vector().local_size()
+        # Set data manager
+        self.material.set_data_manager(self.ngauss)
+        # dummy function used for computing global quantity
+        self._dummy_function = Function(self.W0)
+
+        if self.material.hypothesis == mgis_bv.Hypothesis.Tridimensional:
+            assert self.u.geometric_dimension()==3, "Conflicting geometric dimension and material hypothesis"
+        else:
+            assert self.u.geometric_dimension()==2, "Conflicting geometric dimension and material hypothesis"
+
+        if bcs is None:
+            self.bcs = []
+        else:
+            self.bcs = bcs
[…]
