+++ bindings/python/mgis/fenics/CMakeLists.txt
@@ -5,3 +5,7 @@ function(mgis_fenics_python_script file)
 endfunction(mgis_fenics_python_script)
 
 mgis_fenics_python_script(__init__.py)
+mgis_fenics_python_script(gradient_flux.py)
+mgis_fenics_python_script(utils.py)
+mgis_fenics_python_script(nonlinear_material.py)
+mgis_fenics_python_script(nonlinear_problem.py)

[…]
