### 2020-04-15 ee7bdf9
Fixed CMakeLists for pandoc
