+++ docs/web/mgis_fenics_multiphase_model.md
@@ -1,0 +1,337 @@
+$\newcommand{\bsig}{\boldsymbol{\sigma}}
+\newcommand{\beps}{\boldsymbol{\varepsilon}}
+\newcommand{\bkappa}{\boldsymbol{\kappa}}
+\newcommand{\DD}{\mathbb{D}}
+\newcommand{\CC}{\mathbb{C}}
+\newcommand{\bI}{\boldsymbol{I}}
+\newcommand{\bV}{\boldsymbol{V}}
+\newcommand{\bU}{\boldsymbol{U}}
+\newcommand{\bt}{\boldsymbol{n}}
+\newcommand{\bn}{\boldsymbol{t}}
+\newcommand{\T}{{}^\text{T}} 
+\newcommand{\div}{\operatorname{div}}
+\newcommand{\jump}[1]{[\![#1]\!]}
+\newcommand{\avg}[1]{\left\langle#1\right\rangle}$
+
+
+# Multiphase model for fiber-reinforced materials
+
+
+This demo explores the full capability of MFront's recent extension to handling generalized behaviours, namely involving different pairs of fluxes and gradients, as well as FEniCS versatility to solve generic PDEs.
+
+This will be illustrated on a generalized continuum model called *multiphase model* in the context of fiber-reinforced materials <cite data-cite="bleyer2018multiphase">(Bleyer, 2018)</cite>.
+
+## A quick primer on the multiphase model
+
+The multiphase model is a higher-grade (i.e. with enhanced kinematics) generalized continuum which represents a biphasic material (the main application being fiber-reinforced materials) by a superposition of two phases (say *matrix* and *fiber* phases), each possessing their own kinematics and in mutual interaction. Each phase kinematics is described by a displacement field $\bU^1$ and $\bU^2$ for the matrix and fiber phase respectively.
+
+In the present case, each phase is a standard Cauchy continuum but other variants, including fiber bending effects for instance, exist, see <cite data-cite="debuhan2017elastic">(de Buhan et al., 2017)</cite> for more details.
+
+* **Generalized strains**: infinitesimal strain in each phase $\beps^j=\nabla^s \bU^j$ and relative displacement between both phases $\bV = \bU^2-\bU^1$
+
+* **Generalized stresses**: partial stress $\bsig^j$ in both phases and interaction force $\bI$ (exerted by phase 2 over phase 1)
+
+* **Equilibrium equations**:
+\begin{align}
+\div \bsig^1 +\rho_1F + \bI &=  0 \label{multiphase-eq1}\\
+\div \bsig^2 + \rho_2F - \bI &= 0 \label{multiphase-eq2}
+\end{align}
+
+* **Traction boundary conditions**:
+\begin{align}
+\bsig^1\cdot\bn &= \bt^1 \text{ on }\partial \Omega_{T} \\
+\bsig^2\cdot\bn &= \bt^2 \text{ on } \partial \Omega_{T} 
+\end{align}
+
+* **Deformation work density**:
+\begin{equation}
+w_{\text{def}} = \bsig^1:\beps^1 + \bsig^2:\beps^2+ \bI\cdot\bV
+\end{equation}
+
+* General form of the **constitutive equations** based on a convex free energy density $\psi$:
+\begin{align}
[…]
