### 2020-04-16 3b58e89
Fixes in nonlinear problem
