+++ docs/web/mgis_fenics_multiphase_model.md
@@ -223,9 +223,10 @@ In such a problem, both phase strains are uniform along the vertical direction s
 
 
 ```python
+%matplotlib notebook
 import matplotlib.pyplot as plt
 from dolfin import *
-import mfront_wrapper as mf
+import mgis.fenics as mf
 import numpy as np
 from ufl import diag
 
@@ -328,10 +329,9 @@ plt.show()
 ```
 
 
-![png](multiphase_model_files/multiphase_model_6_0.png)
+    <IPython.core.display.Javascript object>
 
 
 
-```python
+<img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAwAAAAJACAYAAAA6rgFWAAAgAElEQVR4nOzde5QcV30v+i0/sQ24DMGagAnQwElMAuZUQpIbVswJ6pBgLbgX3JfkkOSYgN08DjHnBJ8OAZ8IRgHRYREnDsTuQG6MwwnYnYFBPGTSyGNkjF7TeoyERpamHcmxNRYauz3vmZru/t4/Srunu6v6UbsfVbXr+1mrlq2uftTMb2pmf6v2Q4CIiIiIiCJD+H0AREREREQ0OAwAREREREQRwgBARERERBQhDABERERERBHCAEBEREREFCEMAEREREREEcIAQEREREQUIQwAREREREQRwgBARERERBQhDABERERERBHCAEBEREREFCEMAEREREREEcIAQEREREQUIQwAREREREQRwgBARERERBQhDABERERERBHCAEBEREREFCEMAEREREREEcIAQEREREQUIQwAREREREQRwgBARERERBQhDABERERERBHCAEBEREREFCEMAEREREREEcIAQEREREQUIQwAREREREQRwgBARERERBQhDABERERERBHCAEBEREREFCEMAEREREREEcIAQEREREQUIQwAREREREQRwgBARERERBQhDABERERERBHCAEBEREREFCEMAEREREREEcIAQEREREQUIQwAREREREQRwgBARERERBQhDABERERERBHCAEBEREREFCEMAEREREREEcIAQEREREQUIQwAREREREQRwgBARERERBQhDABERERERBHCAEBEREREFCEMAEREREREEcIAQEREREQUIQwAREREREQRwgBARERERBQhDABERERERBHCAEBEREREFCEMAEREREREEcIAQEREREQUIQwAREREREQRwgBARERERBQhDABERERERBHCAEBEREREFCGhDgC5XA7JZBLpdLr6X1X5fB6JRMLTa9LpNLLZrPJnEhERERENWmgDQDabhWmadY+l02nPjXgAyGQyEEIgHo93/JpisQghBDKZjOfPIyIiIiLyS2gDgGEYyOVyro93elU+kUggHo8jk8nAMAxPASCZTDIAEBEREVHohDIAZLNZCOF+6IlEQukugJcAkMvlk
[…]
