### 2020-04-16 3c88145
Transient heat equation is now OK, must still suppress previous
