+++ bindings/python/mgis/fenics/nonlinear_problem.py
@@ -91,8 +91,6 @@ class AbstractNonlinearProblem:
         self.gradients = dict.fromkeys(self.material.get_gradient_names(), None)
         self.fluxes = {}
 
-        self.initialize_internal_state_variables()
-
     def automatic_registration(self):
         """
         Performs automatic registration of predefined Gradients or external state variables.
@@ -131,10 +129,10 @@ class AbstractNonlinearProblem:
         """Computes derivative of residual"""
         # derivatives of fluxes
         self.tangent_a = sum([derivative(self.residual, f.function, t*dg.variation(self.du))
-                      for f in self.fluxes.values()  for (t, dg) in zip(f.tangent_blocks.values(), f.gradients)])
+                      for f in self.fluxes.values()  for (t, dg) in zip(f.tangent_blocks.values(), f.variables)])
         # derivatives of internal state variables
-        self.tangent_a += sum([derivative(self.residual, f.function, t*dg.variation(self.du))
-                      for f in self.fluxes.values() for (t, dg) in zip(f.tangent_blocks.values(), f.gradients)])
+        self.tangent_a += sum([derivative(self.residual, s.function, t*dg.variation(self.du))
+                      for s in self.state_variables["internal"].values() for (t, dg) in zip(s.tangent_blocks.values(), s.variables)])
         # derivatives of variable u
         self.tangent_a += derivative(self.residual, self.u, self.du)
 
@@ -215,11 +213,44 @@ class AbstractNonlinearProblem:
         # tangent matrix quadrature space
         self.WCt = FunctionSpace(self.mesh, Wce)
 
+    def initialize_fluxes(self):
+        for (f, size) in zip(self.material.get_flux_names(), self.material.get_flux_sizes()):
+            flux_gradients = []
+            for t in self.material.get_tangent_block_names():
+                if t[0]==f:
+                    try:
+                        flux_gradients.append(self.gradients[t[1]])
+                    except:
+                        value = self.state_variables["external"].get(t[1], None)
+                        if value is not None and isinstance(value, Var):
+                                flux_gradients.append(value)
+                        else:
+                            raise ValueError("'{}' could not be associated with a registered gradient or state variable.".format(t[1]))
+            flux = Flux(f, size, flux_gradients)
+            flux.initialize_functions(self.mesh, self.quadrature_degree)
+            self.fluxes.update({f: flux})
+
     def initialize_internal_state_variables(self):
         for (s, size) in zip(self.material.get_internal_state_variable_names(), self.material.get_internal_state_variable_sizes()):
             We = get_quadrature_element(self.mesh.ufl_cell(), self.quadrature_degree, size)
             W = FunctionSpace(self.mesh, We)
             self.state_variables["internal"][s] = Function(W, name=s)
+
+        for (s, size) in zip(self.material.get_internal_state_variable_names(), self.material.get_internal_state_variable_sizes()):
+            state_var_gradients = []
+            for t in self.material.get_tangent_block_names():
+                if t[0]==s:
+                    try:
+                        state_var_gradients.append(self.gradients[t[1]])
+                    except:
+                        value = self.state_variables["external"].get(t[1], None)
+                        if value is not None and isinstance(value, Var):
+                                state_var_gradients.append(value)
+                        else:
+                            raise ValueError("'{}' could not be associated with a registered gradient or state variable.".format(t[1]))
+            state_variable = InternalStateVariable(s, size, state_var_gradients)
+            state_variable.initialize_functions(self.mesh, self.quadrature_degree)
+            self.state_variables["internal"][s] = state_variable
 
     def initialize_fields(self):
         """
[…]
