+++ bindings/python/mgis/fenics/gradient_flux.py
@@ -79,12 +79,11 @@ class Var(Gradient):
     def __init__(self, variable, name):
         return Gradient.__init__(self, variable, variable, name)
 
-
-class Flux:
-    def __init__(self, name, shape, gradients=[]):
+class QuadratureFunction:
+    def __init__(self, name, shape, variables=[]):
         self.shape = shape
         self.name = name
-        self.gradients = gradients
+        self.variables = variables
 
     def initialize_functions(self, mesh, quadrature_degree):
         We = get_quadrature_element(mesh.ufl_cell(), quadrature_degree, self.shape)
@@ -92,10 +91,10 @@ class Flux:
         self.function = Function(W, name=self.name)
         values = [Function(FunctionSpace(mesh,
                           get_quadrature_element(mesh.ufl_cell(),
-                          quadrature_degree, dim=(self.shape, g.shape))),
-                           name="d{}_d{}".format(self.name, g.name))
-                          for g in self.gradients]
-        keys = [g.name for g in self.gradients]
+                          quadrature_degree, dim=(self.shape, v.shape))),
+                           name="d{}_d{}".format(self.name, v.name))
+                          for v in self.variables]
+        keys = [v.name for v in self.variables]
         self.tangent_blocks = dict(zip(keys, values))
 
     def update(self, x):
@@ -108,3 +107,36 @@ class Flux:
             return self.previous
         except:
             raise ValueError("Function must be initialized first.")
+
+class Flux(QuadratureFunction):
+    pass
+
+class InternalStateVariable(QuadratureFunction):
+    pass
+    # def __init__(self, name, shape, gradients=[]):
+    #     self.shape = shape
+    #     self.name = name
+    #     self.gradients = gradients
+
+    # def initialize_functions(self, mesh, quadrature_degree):
+    #     We = get_quadrature_element(mesh.ufl_cell(), quadrature_degree, self.shape)
+    #     W = FunctionSpace(mesh, We)
+    #     self.function = Function(W, name=self.name)
+    #     values = [Function(FunctionSpace(mesh,
+    #                       get_quadrature_element(mesh.ufl_cell(),
+    #                       quadrature_degree, dim=(self.shape, g.shape))),
+    #                        name="d{}_d{}".format(self.name, g.name))
+    #                       for g in self.gradients]
+    #     keys = [g.name for g in self.gradients]
+    #     self.tangent_blocks = dict(zip(keys, values))
+
+    # def update(self, x):
+    #     self.function.vector().set_local(x)
+
+    # def previous(self):
+    #     try:
+    #         self.previous = Function(self.function.function_space(), name=self.name+"_previous")
+    #         self.previous.assign(self.function)
+    #         return self.previous
+    #     except:
+            # raise ValueError("Function must be initialized first.")
