+++ bindings/python/mgis/fenics/gradient_flux.py
@@ -80,12 +80,15 @@ class Var(Gradient):
         return Gradient.__init__(self, variable, variable, name)
 
 class QuadratureFunction:
+    """An abstract class for Flux and InternalStateVariables"""
     def __init__(self, name, shape, variables=[]):
         self.shape = shape
         self.name = name
         self.variables = variables
 
     def initialize_functions(self, mesh, quadrature_degree):
+        self.quadrature_degree = quadrature_degree
+        self.mesh = mesh
         We = get_quadrature_element(mesh.ufl_cell(), quadrature_degree, self.shape)
         W = FunctionSpace(mesh, We)
         self.function = Function(W, name=self.name)
@@ -100,43 +103,29 @@ class QuadratureFunction:
     def update(self, x):
         self.function.vector().set_local(x)
 
-    def previous(self):
-        try:
-            self.previous = Function(self.function.function_space(), name=self.name+"_previous")
-            self.previous.assign(self.function)
-            return self.previous
-        except:
-            raise ValueError("Function must be initialized first.")
+    def project_on(self, space, degree):
+        """
+        Returns the projection on a standard CG/DG space
+
+        Parameters
+        ----------
+
+        space: str
+            FunctionSpace type ("CG", "DG",...)
+        degree: int
+            FunctionSpace degree
+        """
+        if self.shape == 1:
+            V = FunctionSpace(self.mesh, space, degree)
+        else:
+            V = VectorFunctionSpace(self.mesh, space, degree, dim=self.shape)
+        v = Function(V, name=self.name)
+        v.assign(project(self.function, V,
+                         form_compiler_parameters={"quadrature_degree": self.quadrature_degree}))
+        return v
 
 class Flux(QuadratureFunction):
     pass
 
 class InternalStateVariable(QuadratureFunction):
     pass
-    # def __init__(self, name, shape, gradients=[]):
-    #     self.shape = shape
-    #     self.name = name
-    #     self.gradients = gradients
-
-    # def initialize_functions(self, mesh, quadrature_degree):
-    #     We = get_quadrature_element(mesh.ufl_cell(), quadrature_degree, self.shape)
-    #     W = FunctionSpace(mesh, We)
-    #     self.function = Function(W, name=self.name)
-    #     values = [Function(FunctionSpace(mesh,
-    #                       get_quadrature_element(mesh.ufl_cell(),
-    #                       quadrature_degree, dim=(self.shape, g.shape))),
-    #                        name="d{}_d{}".format(self.name, g.name))
-    #                       for g in self.gradients]
-    #     keys = [g.name for g in self.gradients]
-    #     self.tangent_blocks = dict(zip(keys, values))
-
-    # def update(self, x):
-    #     self.function.vector().set_local(x)
-
-    # def previous(self):
-    #     try:
-    #         self.previous = Function(self.function.function_space(), name=self.name+"_previous")
-    #         self.previous.assign(self.function)
-    #         return self.previous
-    #     except:
-            # raise ValueError("Function must be initialized first.")
