### 2020-04-17 bc47fb6
Renamed some methods in nonlinear problem and removed unused code
