+++ docs/web/mgis_fenics.md
@@ -51,7 +51,7 @@ This class handles the definition and resolution of a nonlinear problem associat
 
 * `material`: the associated `MFrontNonlinearMaterial`
 
-* `state_variables`: a dictionary of internal and external state variables represented as `dolfin` objects
+* `state_variables`: a dictionary of internal and external state variables. External state variables are represented as `dolfin` objects whereas internal state variables are represented as [Quadrature functions](#flux-and-gradient-objects).
 
 * `gradients`: a dictionary of `Gradient` objects
 
@@ -60,6 +60,26 @@ This class handles the definition and resolution of a nonlinear problem associat
 * `residual`: the nonlinear residual $F(u)$
 
 * `tangent_form`: the tangent bilinear form associated with the residual
+
+* `solver`: the non-linear solver (default is `NewtonSolver`)
+
+Its main methods are:
+
+* `register_gradient`: registers a MFront gradient with a UFL expression (see [the registration concept](#the-registration-concept)).
+
+* `register_external_state_variable`: registers a MFront external state variable with a UFL expression (see [the registration concept](#the-registration-concept)).
+
+* `set_loading`: defines the external forces linear form $L(v)$ in the default residual expression 
+
+* `initialize`: initializes the functions associated with gradients, fluxes, external and internal state variables objects and the corresponding tangent blocks. All gradients and external state variables must have been registered first. [Automatic registration](#the-registration-concept) is performed at the beginning of the method call. This method has to be called once before calling `solve`.
+
+* `update_constitutive_law`: performs the consitutive law update. Currently, this method is called for all quadrature points before assembling the residual or tangent form. Ideally, this should be done locally during the assembly (see [the module current limitations](#current-limitations)).
+
+* `solve`: solves the associated non-linear problem using `solver`
+
+* `get_flux`: returns the function associated with a flux
+
+* `get_state_variable`: returns the function associated with an internal state variable
 
 ### Residual and tangent bilinear form
 By default, the nonlinear residual is assumed to take the following form: Find $u\in V$ such that:
@@ -76,12 +96,12 @@ where the $\bsig_i(u)$ are a set of **fluxes** as defined in the MFront behaviou
 
 This generic form is suitable for most quasi-static mechanical problems but it does not necessarily encompass all possible situations. In particular, evolution equations such as transient heat transfer cannot be expressed in this form. However, this is not a limitation since the residual can also be redefined explicitly by the user (see [Transient non-linear heat equation](demos/transient_nonlinear_heat_equation.html)).
 
-Either for the default or a user-defined one, the tangent operator associated with the residual is computed automatically using either the UFL symbolic derivation `ufl.derivative` for simple terms or using tangent operator blocks defined in the MFront behaviour for the nonlinear fluxes. More precisely, the variation of each flux is given by:
+Either for the default or a user-defined one, the tangent operator associated with the residual is computed automatically using either the UFL symbolic derivation `ufl.derivative` for simple terms (e.g. involving the unknown field $u$) or using tangent operator blocks defined in the MFront behaviour for the nonlinear fluxes and internal state variables. More precisely, the variation of each flux is given by:
 
 $$
 \dfrac{\partial \bsig_i}{\partial u} = \sum_{j\in \text{blocks}(i)} \TT^{\bsig_i}_{\bg_j}\cdot \delta \bg_j(u)
 $$
-where the flux $\bsig_i$ is assumed to depend on the gradients $\bg_j$ for $j\in \text{blocks}(i)$ (which at least contains $i$ itself in general) and $\TT^{\bsig_i}_{\bg_j} = \dfrac{\partial \bsig_i}{\partial \bg_j}$ is the tangent operator associated with the corresponding block.
[…]
