### 2020-04-17 d0f1d8e
Add automatic call to mfront
