### 2020-04-17 985e687
Add automatic mfront compilation
