+++ bindings/python/mgis/fenics/nonlinear_material.py
@@ -11,6 +11,7 @@ import mgis.behaviour as mgis_bv
 from .gradient_flux import Var
 import dolfin
 import subprocess
+import os
 
 mgis_hypothesis = {"plane_strain": mgis_bv.Hypothesis.PlaneStrain,
                    "plane_stress": mgis_bv.Hypothesis.PlaneStress,
@@ -48,25 +49,29 @@ class MFrontNonlinearMaterial:
         self.hypothesis = mgis_hypothesis[hypothesis]
         self.material_properties = material_properties
         # Loading the behaviour
-        self.is_finite_strain = mgis_bv.isStandardFiniteStrainBehaviour(self.path, self.name)
         try:
             self.load_behaviour()
         except:
-            print("Behaviour '{}' has not been found '{}' in '{}'.".format(self.name, self.path))
-            print("Attempting to compile '{}.mfront'...".format(self.name))
+            cwd = os.getcwd()
+            install_path = "/".join(path.split("/")[:-2])+"/"
+            os.chdir(install_path)
+            print("Behaviour '{}' has not been found in '{}'.".format(self.name, self.path))
+            print("Attempting to compile '{}.mfront' in {}...".format(self.name, install_path))
             subprocess.run(["mfront", "--obuild", "--interface=generic", self.name+".mfront"])
+            os.chdir(cwd)
             self.load_behaviour()
         self.update_parameters(parameters)
 
     def load_behaviour(self):
-            if self.is_finite_strain:
-                # finite strain options
-                bopts = mgis_bv.FiniteStrainBehaviourOptions()
-                bopts.stress_measure = mgis_bv.FiniteStrainBehaviourOptionsStressMeasure.PK1
-                bopts.tangent_operator = mgis_bv.FiniteStrainBehaviourOptionsTangentOperator.DPK1_DF
-                self.behaviour = mgis_bv.load(bopts, self.path, self.name, self.hypothesis)
-            else:
-                self.behaviour = mgis_bv.load(self.path, self.name, self.hypothesis)
+        self.is_finite_strain = mgis_bv.isStandardFiniteStrainBehaviour(self.path, self.name)
+        if self.is_finite_strain:
+            # finite strain options
+            bopts = mgis_bv.FiniteStrainBehaviourOptions()
+            bopts.stress_measure = mgis_bv.FiniteStrainBehaviourOptionsStressMeasure.PK1
+            bopts.tangent_operator = mgis_bv.FiniteStrainBehaviourOptionsTangentOperator.DPK1_DF
+            self.behaviour = mgis_bv.load(bopts, self.path, self.name, self.hypothesis)
+        else:
+            self.behaviour = mgis_bv.load(self.path, self.name, self.hypothesis)
 
     def set_data_manager(self, ngauss):
         # Setting the material data manager
