### 2020-04-18 d060254
Moved flux and isv initialization
