### 2020-04-18 0f56704
Fixed issue in finite strain
