### 2020-04-20 0021f08
Fixed shapes of tangent operators
