+++ bindings/python/mgis/fenics/nonlinear_problem.py
@@ -130,10 +130,18 @@ class AbstractNonlinearProblem:
         """Computes derivative of residual"""
         # derivatives of fluxes
         self.tangent_form = sum([derivative(self.residual, f.function, t*dg.variation(self.du))
-                      for f in self.fluxes.values()  for (t, dg) in zip(f.tangent_blocks.values(), f.variables)])
+                      for f in self.fluxes.values()
+                      for (t, dg) in zip(f.tangent_blocks.values(), f.variables)])
         # derivatives of internal state variables
-        self.tangent_form += sum([derivative(self.residual, s.function, t*dg.variation(self.du))
-                      for s in self.state_variables["internal"].values() for (t, dg) in zip(s.tangent_blocks.values(), s.variables)])
+        for s in self.state_variables["internal"].values():
+            for (t, dg) in zip(s.tangent_blocks.values(), s.variables):
+                tdg = t*dg.variation(self.du)
+                if len(shape(t)) >0 and shape(t)[0] == 1:
+                    tdg = tdg[0]
+                # print(s.name, t.name())
+                # print(shape(t), shape(dg.variation(self.du)))
+                # derivative(self.residual, ss, tdg)
+                self.tangent_form += derivative(self.residual, s.function, tdg)
         # derivatives of variable u
         self.tangent_form += derivative(self.residual, self.u, self.du)
 
@@ -181,7 +189,7 @@ class AbstractNonlinearProblem:
         elif type(expression) == float:
             self.state_variables["external"].update({name: Constant(expression)})
         else:
-            self.state_variables["external"].update({name: Var(expression, name)})
+            self.state_variables["external"].update({name: Var(self.u, expression, name)})
 
     def set_loading(self, Fext):
         """
