### 2020-04-20 d2aa30a
More fix on tangent computation for mixed space
