### 2020-04-21 34e4f72
Error message for automatic registration using mixed function spaces
