+++ bindings/python/mgis/fenics/nonlinear_problem.py
@@ -100,15 +100,15 @@ class AbstractNonlinearProblem:
             if value is None:
                 case = predefined_gradients.get(name, None)
                 if case is not None:
-                    if len(split(self.u)) == 1:
+                    if self.u.ufl_element().family() == "Mixed":
+                        error_msg = "Automatic registration cannot be used with mixed function spaces.\n"
+                        error_msg += "Gradient '{}' must be registered explicitly.".format(name)
+                        raise NotImplementedError(error_msg)
+                    else:
                         if self.axisymmetric:
                             var = (self._r, self.u)
                         else:
                             var = (self.u,)
-                    else:
-                        error_msg = "Automatic registration cannot be used with mixed function spaces.\n"
-                        error_msg += "Gradient '{}' must be registered explicitly.".format(name)
-                        raise NotImplementedError(error_msg)
                     expr = case[self.material.hypothesis](*var)
                     self.register_gradient(name, expr)
                     print("Automatic registration of '{}' as {}.\n".format(name, str(expr)))
