+++ demos/mgis_fenics/supplementary_files/references.bib
@@ -1,0 +1,105 @@
+@article{bleyer2018multiphase,
+  title={Multiphase continuum models for fiber-reinforced materials},
+  author={Bleyer, Jeremy},
+  journal={Journal of the Mechanics and Physics of Solids},
+  volume={121},
+  pages={198--233},
+  year={2018},
+  publisher={Elsevier}
+}
+
+@@book{debuhan2017elastic,
+  title={Elastic, Plastic and Yield Design of Reinforced Structures},
+  author={De Buhan, Patrick and Bleyer, J{\'e}r{\'e}my and Hassen, Ghazi},
+  year={2017},
+  publisher={Elsevier}
+}
+
+@article{miehe_anisotropic_2002,
+	title = {Anisotropic additive plasticity in the logarithmic strain space: modular kinematic formulation and implementation based on incremental minimization principles for standard materials},
+	volume = {191},
+	issn = {00457825},
+	shorttitle = {Anisotropic additive plasticity in the logarithmic strain space},
+	url = {http://linkinghub.elsevier.com/retrieve/pii/S0045782502004383},
+	doi = {10.1016/S0045-7825(02)00438-3},
+	abstract = {The paper presents a modular formulation and computational implementation of a class of anisotropic plasticity models at ﬁnite strains based on incremental minimization principles. The modular kinematic setting consists of a constitutive model in the logarithmic strain space that is framed by a purely geometric pre- and postprocessing. On the theoretical side, the point of departure is an a priori six-dimensional approach to ﬁnite plasticity based on the notion of a plastic metric. In a ﬁrst step, a geometric preprocessor deﬁnes a total and a plastic logarithmic strain measure obtained from the current and the plastic metrics, respectively. In a second step, these strains enter in an additive format a constitutive model of anisotropic plasticity that may have a structure identical to the geometrically linear theory. The model deﬁnes the stresses and consistent tangents work-conjugate to the logarithmic strain measure. In a third step these objects of the logarithmic space are then mapped back to nominal, Lagrangian or Eulerian objects by a geometric postprocessor. This geometric three-step-approach deﬁnes a broad class of anisotropic models of ﬁnite plasticity directly related to counterparts of the geometrically linear theory. It is speciﬁed to a model problem of anisotropic metal plasticity. On the computational side we develop an incremental variational formulation of the above outlined constitutive structure where a quasi-hyperelastic stress potential is obtained from a local constitutive minimization problem with respect to the internal variables. It is shown that this minimization problem is exclusively restricted to the logarithmic strain space in a structure identical to the small-strain theory. The minimization problem determines the internal state of the material for ﬁnite increments of time. We develop a discrete formulation in terms of just one scalar parameter for the amount of incremental ﬂow. The existence of the incremental stress potential provides a natural basis for the deﬁnition of the geometric postprocessor based on function evaluations. Furthermore, the global initialboundary-value-problem of the elastic–plastic solid appears in the incremental setting as an energy minimization problem. Numerical examples show that the results obtained are surprisingly close to those obtained by a reference framework of multiplicative plasticity.},
+	language = {en},
+	number = {47-48},
+	urldate = {2019-02-04},
+	journal = {Computer Methods in Applied Mechanics and Engineering},
+	author = {Miehe, C. and Apel, N. and Lambrecht, M.},
+	month = nov,
+	year = {2002},
[…]
