+++ docs/web/mgis_fenics_nonlinear_heat_transfer.md
@@ -1,6 +1,6 @@
-# Stationnary non-linear heat transfer
+% Stationnary non-linear heat transfer
 
-## Description of the non-linear constitutive heat transfer law
+# Description of the non-linear constitutive heat transfer law
 
 The thermal material is described by the following non linear Fourier
 Law:
@@ -12,8 +12,7 @@ $$
 where $\mathbf{j}$ is the heat flux and $\mathbf{\nabla} T$ is the
 temperature gradient.
 
-Expression of the thermal conductivity
---------------------------------------
+## Expression of the thermal conductivity
 
 The thermal conductivity is assumed to be given by:
 
@@ -24,8 +23,7 @@ $$
 This expression accounts for the phononic contribution to the thermal
 conductivity.
 
-Derivatives
------------
+## Derivatives
 
 As discussed below, the consistent linearisation of the heat transfer
 equilibrium requires to compute:
@@ -45,11 +43,9 @@ equilibrium requires to compute:
       {\displaystyle \frac{\displaystyle \partial \mathbf{j}}{\displaystyle \partial T}}=-{\displaystyle \frac{\displaystyle \partial k\left(T\right)}{\displaystyle \partial T}}\,\mathbf{\nabla} T=B\,k^{2}\,\mathbf{\nabla} T
     $$
 
-`MFront`’ implementation
-========================
+# `MFront`’ implementation
 
-Choice of the the domain specific language
-------------------------------------------
+## Choice of the the domain specific language
 
 Every `MFront` file is handled by a domain specific language (DSL), which
 aims at providing the most suitable abstraction for a particular choice
@@ -68,8 +64,7 @@ The `DefaultGenericBehaviour` is the most suitable choice:
 @DSL DefaultGenericBehaviour;
 ```
 
-Some metadata
--------------
+## Some metadata
 
 The following lines define the name of the behaviour, the name of the
 author and the date of its writing:
@@ -80,8 +75,7 @@ author and the date of its writing:
 @Date 15/02/2019;
 ```
 
-Gradients and fluxes
---------------------
+## Gradients and fluxes
 
 Generic behaviours relate pairs of gradients and fluxes. Gradients and
 fluxes are declared independently but the first declared gradient is
@@ -112,8 +106,7 @@ j.setGlossaryName("HeatFlux");
 In the following code blocks, `j` will be the heat flux at the end of
 the time step.
 
-Tangent operator blocks
------------------------
+## Tangent operator blocks
 
 By default, the derivatives of the gradients with respect to the fluxes
 are declared. Thus the variable `∂j∕∂Δ∇T` is automatically declared.
@@ -127,8 +120,7 @@ with respect to the temperature at the end of the time step).
 @AdditionalTangentOperatorBlock ∂j∕∂ΔT;
 ```
 
-Parameters
-----------
+## Parameters
 
 The `A` and `B` coefficients that appears in the definition of the
 thermal conductivity are declared as parameters:
@@ -142,8 +134,7 @@ Parameters are stored globally and can be modified from the calling
 solver or from `python` in the case of the coupling with `FEniCS`
 discussed below.
 
-Local variable
---------------
+## Local variable
 
 A local variable is accessible in each code blocks.
 
@@ -155,8 +146,7 @@ and to reuse this value when computing the tangent operator.
 @LocalVariable thermalconductivity k;
 ```
 
-Integration of the behaviour
-----------------------------
+## Integration of the behaviour
 
 The behaviour integration is straightforward: one starts to compute the
 temperature at the end of the time step, then we compute the thermal
@@ -174,8 +164,7 @@ temperature gradient (at the end of the time step).
 } // end of @Integrator
 ```
 
-Tangent operator
-----------------
[…]
