+++ docs/web/bibliography.bib
@@ -7,6 +7,24 @@
     place={Oulu}
 }
 
+@article{bleyer2018multiphase,
+  title={Multiphase continuum models for fiber-reinforced materials},
+  author={Bleyer, Jeremy},
+  journal={Journal of the Mechanics and Physics of Solids},
+  volume={121},
+  pages={198--233},
+  year={2018},
+  publisher={Elsevier}
+}
+
+@book{debuhan2017elastic,
+  title={Elastic, Plastic and Yield Design of Reinforced Structures},
+  author={De Buhan, Patrick and Bleyer, J{\'e}r{\'e}my and Hassen, Ghazi},
+  year={2017},
+  publisher={Elsevier}
+}
+
+
 @misc{bleyer_fenics_2019,
 	title = {{FEniCS} and {MFront} for complex non linear solid mechanics simulation},
 	url = {https://www.researchgate.net/publication/333089270_FEniCS_and_MFront_for_complex_non_linear_solid_mechanics_simulation},

[…]
