### 2020-04-22 6c19c96
Add processing sources script
