### 2020-04-22 #2003 @IgorBaratta
I finished my review. All boxes are now checked.
I was able to test it with the FEniCS and JuliaFEM libraries.
Also, the documentation is much better now.

S now, I recommend it for publication.
