### 2020-04-22 #2003 @arfon
@thelfer - I made a bunch of small changes to the paper in https://github.com/thelfer/MFrontGenericInterfaceSupport/pull/40 which are mostly designed to fix citation formatting.

Could you also review the suggested DOIs from Whedon [above](https://github.com/openjournals/joss-reviews/issues/2003#issuecomment-617948170) and update your BibTeX file where appropriate?

Finally, could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
