### 2020-04-22 ae74c7b
Adds missing dois to the bibliography
