+++ docs/papers/joss/bibliography.bib
@@ -37,7 +37,7 @@
 @article{bezanson_julia:_2017,
 	title = {Julia: A fresh approach to numerical computing},
 	volume = {59},
-	url = {https://doi.org/10.1137/141000671},
+	doi = {10.1137/141000671},
 	pages = {65--98},
 	number = {1},
 	journaltitle = {{SIAM} review},
@@ -54,7 +54,7 @@
     volume={52},
     number={3},
     pages={160--176},
-    url={https://doi.org/10.23998/rm.76193}
+    doi={10.23998/rm.76193}
 }
 
 @article{rapo_pipe_2019,
@@ -66,7 +66,7 @@
     volume={52},
     number={3},
     pages={177--191},
-    url={https://doi.org/10.23998/rm.76259}
+    doi={10.23998/rm.76259}
 }
 
 @article{aho_introduction_2019,
@@ -78,7 +78,7 @@
     volume={52},
     number={3},
     pages={148--159},
-    url={https://doi.org/10.23998/rm.75103}
+    doi={10.23998/rm.75103}
 }
 
 
@@ -91,7 +91,7 @@
   year={2017},
   pages={300--303},
   note={open access},
-  url={https://doi.org/10.23998/rm.65040}
+  doi={10.23998/rm.65040}
 }
 
 @article{Nagel2016,
@@ -392,14 +392,13 @@ Untersuchungen zu homogenen Ersatzsystemen}},
   publisher={in Rakenteiden mekaniikan seura Ry},
   pages = {229--233},
   note={open access},
-  url={https://doi.org/10.23998/rm.64224},
-  NOSHOWurl="https://www.overleaf.com/read/sdfzvdxbcxwm"
+  doi={10.23998/rm.64224},
 }
 
 @article{frondelius_juliafem_2017,
   title={Julia{FEM} - open source solver for both industrial and academia usage},
   volume={50},
-  url= {https://doi.org/10.23998/rm.64224},
+  doi={10.23998/rm.64224},
   number={3},
   journal={Rakenteiden Mekaniikka},
   author={Frondelius, Tero and Aho, Jukka},
@@ -415,7 +414,7 @@ Untersuchungen zu homogenen Ersatzsystemen}},
   number=1,
   year=2018,
   pages={36-54},
-  url={https://doi.org/10.23998/rm.69026},
+  doi={10.23998/rm.69026},
   note={open access}
 }
 
