### 2020-04-22 2b3d2f1
add the documentation of the future mgis.fenics python module
