+++ docs/web/mgis_fenics_nonlinear_heat_transfer.md
@@ -1,0 +1,368 @@
+% Stationnary non-linear heat transfer
+% Jérémy Bleyer, Thomas Helfer
+% 17/04/2020
+
+# Description of the non-linear constitutive heat transfer law
+
+The thermal material is described by the following non linear Fourier
+Law:
+
+$$
+\mathbf{j}=-k\left(T\right)\,\mathbf{\nabla} T
+$$
+
+where $\mathbf{j}$ is the heat flux and $\mathbf{\nabla} T$ is the
+temperature gradient.
+
+## Expression of the thermal conductivity
+
+The thermal conductivity is assumed to be given by:
+
+$$
+k\left(T\right)={\displaystyle \frac{\displaystyle 1}{\displaystyle A+B\,T}}
+$$
+
+This expression accounts for the phononic contribution to the thermal
+conductivity.
+
+## Derivatives
+
+As discussed below, the consistent linearisation of the heat transfer
+equilibrium requires to compute:
+
+-   the derivative
+    ${\displaystyle \frac{\displaystyle \partial \mathbf{j}}{\displaystyle \partial \mathbf{\nabla} T}}$
+    of the heat flux with respect to the temperature gradient.
+    ${\displaystyle \frac{\displaystyle \partial \mathbf{j}}{\displaystyle \partial \mathbf{\nabla} T}}$
+    is given by: $$
+      {\displaystyle \frac{\displaystyle \partial \mathbf{j}}{\displaystyle \partial \mathbf{\nabla} T}}=-k\left(T\right)\,\matrix{I}
+    $$
+-   the derivative
+    ${\displaystyle \frac{\displaystyle \partial \mathbf{j}}{\displaystyle \partial T}}$
+    of the heat flux with respect to the temperature.
+    ${\displaystyle \frac{\displaystyle \partial \mathbf{j}}{\displaystyle \partial T}}$
+    is given by: $$
+      {\displaystyle \frac{\displaystyle \partial \mathbf{j}}{\displaystyle \partial T}}=-{\displaystyle \frac{\displaystyle \partial k\left(T\right)}{\displaystyle \partial T}}\,\mathbf{\nabla} T=B\,k^{2}\,\mathbf{\nabla} T
+    $$
+
+# `MFront`’ implementation
+
+## Choice of the the domain specific language
+
+Every `MFront` file is handled by a domain specific language (DSL), which
+aims at providing the most suitable abstraction for a particular choice
+of behaviour and integration algorithm. See `mfront mfront --list-dsl`
+for a list of the available DSLs.
+
+The name of DSL’s handling generic behaviours ends with
+`GenericBehaviour`. The first part of a DSL’s name is related to the
+integration algorithm used.
+
+In the case of this non linear transfer behaviour, the heat flux is
+explicitly computed from the temperature and the temperature gradient.
+The `DefaultGenericBehaviour` is the most suitable choice:
+
+``` cpp
+@DSL DefaultGenericBehaviour;
+```
+
+## Some metadata
+
+The following lines define the name of the behaviour, the name of the
+author and the date of its writing:
+
+``` cpp
+@Behaviour StationaryHeatTransfer;
+@Author Thomas Helfer;
+@Date 15/02/2019;
+```
+
+## Gradients and fluxes
+
+Generic behaviours relate pairs of gradients and fluxes. Gradients and
+fluxes are declared independently but the first declared gradient is
+assumed to be conjugated with the first declared fluxes and so on…
+
+The temperature gradient is declared as follows (note that Unicode characters are supported):
+
+``` cpp
+@Gradient TemperatureGradient ∇T;
+∇T.setGlossaryName("TemperatureGradient");
+```
+
[…]
