+++ docs/web/mgis_fenics_nonlinear_heat_transfer_3D.md
@@ -1,0 +1,136 @@
+% Stationnary non-linear heat transfer: 3D problem and performance comparisons
+% Jérémy Bleyer, Thomas Helfer
+% 17/04/2020
+
+# Description of the non-linear constitutive heat transfer law
+
+This example is a direct continuation of the [previous 2D example on non-linear heat transfer](nonlinear_heat_transfer.html). The present computations will use the same behaviour `StationaryHeatTransfer.mfront`.
+
+
+![]("img/fuel_rod_solution.png"){width=75%)
+
+
+# FEniCS implementation
+
+We now consider a portion of nuclear fuel rod (Uranium Dioxide $\text{UO}_2$) subject to an external imposed temperature $T_{ext}=1000\text{ K}$ and uniform volumetric heat source $r=300 \text{ MW/m}^3$. From the steady state heat balance equation $\operatorname{div}\mathbf{j} = r$, the variational formulation is now:
+
+\begin{equation}
+F(\widehat{T}) = \int_\Omega \mathbf{j}(T,\nabla T)\cdot\nabla \widehat{T}\,\text{dx} + \int_\Omega r \widehat{T} \,\text{dx}=0 \quad \forall \widehat{T}
+\end{equation}
+
+which fits the general default format of a `MFrontNonlinearProblem`:
+\begin{equation}
+F(\widehat{u}) = \sum_i \int_\Omega \boldsymbol{\sigma}_i(u)\cdot \mathbf{g}_i(\widehat{u})\,\text{dx} -L(\widehat{u}) =0 \quad \forall \widehat{u}
+\end{equation}
+
+where $(\boldsymbol{\sigma}_i,\mathbf{g}_i)$ are pairs of dual flux/gradient and here the external loading form $L$ is given by $-\int_\Omega r \widehat{T} \,\text{dx}$. Compared to the previous example, we just add this source term using the `set_loading` method. Here we use a quadratic interpolation for the temperature field and external temperature is imposed on the surface numbered 12. Finally, we also rely on automatic registration of the gradient and external state variables as explained in the previous demo.
+
+
+```python
+from dolfin import *
+import mgis.fenics as mf
+from time import time
+
+mesh = Mesh()
+with XDMFFile("fuel_rod_mesh.xdmf") as infile:
+    infile.read(mesh)
+mvc = MeshValueCollection("size_t", mesh, 2)
+with XDMFFile("fuel_rod_mf.xdmf") as infile:
+    infile.read(mvc, "facets")
+facets = cpp.mesh.MeshFunctionSizet(mesh, mvc)
+
+V = FunctionSpace(mesh, "CG", 2)
+T = Function(V, name="Temperature")
+T_ = TestFunction(V)
+dT = TrialFunction(V)
+T0 = Constant(300.)
+T.interpolate(T0)
+
+Text = Constant(1e3)
+bc = DirichletBC(V, Text, facets, 12)
+
+r = Constant(3e8)
+
+quad_deg = 2
+material = mf.MFrontNonlinearMaterial("../materials/src/libBehaviour.so",
+                                      "StationaryHeatTransfer")
+problem = mf.MFrontNonlinearProblem(T, material, quadrature_degree=quad_deg, bcs=bc)
+problem.set_loading(-r*T*dx)
+```
+
+The `solve` method computing time is monitored:
+
+
+```python
+tic = time()
+problem.solve(T.vector())
+print("MFront/FEniCS solve time:", time()-tic)
+```
+
+    Automatic registration of 'TemperatureGradient' as grad(Temperature).
+    
+    Automatic registration of 'Temperature' as an external state variable.
+    
+    Calling FFC just-in-time (JIT) compiler, this may take some time.
[…]
