+++ docs/web/CMakeLists.txt
@@ -61,17 +61,24 @@ if(MGIS_HAVE_PANDOC)
   mgis_pandoc_generate_html_page(install)
   mgis_pandoc_generate_html_page(bindings-cxx "--toc" "--toc-depth=3")
   mgis_pandoc_generate_html_page(FEniCSBindings)
+  mgis_pandoc_generate_html_page(mgis_fenics "--toc" "--toc-depth=3")
   mgis_pandoc_generate_html_page(release-notes-1.1 "--toc" "--toc-depth=3")
   mgis_pandoc_generate_html_page(release-notes-1.2 "--toc" "--toc-depth=3")
+  mgis_pandoc_generate_html_page(mgis_fenics_nonlinear_heat_transfer)
+  mgis_pandoc_generate_html_page(mgis_fenics_nonlinear_heat_transfer_3D)
+  mgis_pandoc_generate_html_page(mgis_fenics_small_strain_elastoplasticity)
+  mgis_pandoc_generate_html_page(mgis_fenics_finite_strain_elastoplasticity)
+  mgis_pandoc_generate_html_page(mgis_fenics_multiphase_model)
+  mgis_pandoc_generate_html_page(mgis_fenics_phase_field)
 endif(MGIS_HAVE_PANDOC)
 
 if(MGIS_HAVE_JUPYTER_NBCONVERT)
-  mgis_jupyter_nbconvert_generate_html_page(mgis_fenics_multiphase_model)
+#  mgis_jupyter_nbconvert_generate_html_page(mgis_fenics_multiphase_model)
 endif(MGIS_HAVE_JUPYTER_NBCONVERT)
 
 
 if(MGIS_APPEND_SUFFIX)
-  install(DIRECTORY css
+  install(DIRECTORY css img
     DESTINATION share/doc/mgis-${MGIS_SUFFIX}/web
     COMPONENT website
     PATTERN ".svn" EXCLUDE)
@@ -85,7 +92,7 @@ if(MGIS_APPEND_SUFFIX)
 #    DESTINATION share/doc/mgis-${MGIS_SUFFIX}/web
 #    COMPONENT website)
 else(MGIS_APPEND_SUFFIX)
-  install(DIRECTORY css
+  install(DIRECTORY css img
     DESTINATION share/doc/mgis/web
     COMPONENT website
     PATTERN ".svn" EXCLUDE)

[…]
