### 2020-04-22 5c6ed64
merge with the master branch to prepare a release associated with the publication of the JOSS paper
