+++ docs/papers/joss/paper.md
@@ -18,25 +18,25 @@ authors:
     affiliation: "5, 6"
   - name: Thomas Nagel
     orcid: 0000-0001-8459-4616
-    affiliation: 7
+    affiliation: "7, 8"
   - name: Dmitri Naumov
-    affiliation: 8
+    affiliation: "7, 8"
 affiliations:
-  - name: French Alternative Energies and Atomic Energy Commission
+  - name: CEA, DES, IRESNE, DEC, Cadarache F-13108 Saint-Paul-Lez-Durance, France
     index: 1
   - name: Laboratoire Navier UMR 8205 (École des Ponts ParisTech-IFSTTAR-CNRS)
     index: 2
-  - name: Wärtsilä
+  - name: R&D and Engineering, Wärtsilä, P.O. Box 244, 65101 Vaasa, Finland
     index: 3
-  - name: University of Oulu
+  - name: University of Oulu, Erkki Koiso-Kanttilan katu 1, 90014 Oulu, Finland
     index: 4
-  - name: VTT Technical Research Centre of Finland
+  - name: VTT Technical Research Centre of Finland, Kivimiehentie 3, 02150 Espoo, Finland
     index: 5
-  - name: Aalto University
+  - name: Department of Computer Science, Aalto University, Konemiehentie 2, 02150 Espoo, Finland
     index: 6
-  - name: Technische Universität Bergakademie Freiberg
+  - name: Geotechnical Institute, Technische Universität Bergakademie Freiberg, Gustav-Zeuner-Str. 1, 09599 Freiberg, Germany
     index: 7
-  - name: Helmholtz Centre for Environmental Research -- UFZ
+  - name: Department of Environmental Informatics, Helmholtz Centre for Environmental Research -- UFZ, Permoserstr. 15, 04318 Leipzig, Germany
     index: 8
 date: 8 October 2019
 bibliography: bibliography.bib
@@ -48,89 +48,102 @@ pandoc -f markdown_strict --bibliography=bibliography.bib --filter pandoc-citepr
 
 # Introduction
 
+The ability to easily integrate user-defined constitutive equations
+plays a major role in the versatility of (mechanical) solvers^[The term
+solver emphasizes that the numerical method used to discretize the
+equilibrium equations is not significant in the present context.].
+
 Constitutive equations describe how the internal state variables of a
 material evolve with changing external conditions or mechanical
-loadings. Those state variables can describe many microstructural
+loading. Those state variables can describe many microstructural
 aspects of the material (grain size, dislocation density, hardening
 state, etc.) or be phenomenological in nature (equivalent plastic
 strain). The knowledge of those internal state variables allows the
 computation of local thermodynamic forces which affect the material
 equilibrium at the structural scale.
+Due to the large number of phenomena that can be described
+in this manner, such as plasticity,
+viscoplasticity, or damage, computational mechanics is one of the
+most demanding domains for advanced constitutive equations.
 
 At each time step, the constitutive equations must be integrated to
 obtain the state of the material at the end of the time step. As most
 phenomena are nonlinear, an iterative scheme is required at the
 equilibrium scale to find the local loading of the material: the
-integration of the constitutive equations is thus called several times
-with different estimates of the loading of the material.
-
-Due to the large number of phenomena described (plasticity,
-viscoplasticity, damage, etc.), computational mechanics is one of the
-most demanding domains for advanced constitutive equations.
-
-The ability to easily integrate user-defined constitutive equations
-plays a major role in the versatility of (mechanical) solvers^[The term
-solver emphasizes that the numerical method used to discretize the
-equilibrium equations is not significant.].
+integration of the constitutive equations is thus therefore called several times
+with different estimates of the loading of the material. Algorithmic
+efficiency at the constitutive level is therefore a key aspect for the
+overall efficiency of a code.
 
[…]
