+++ CONTRIBUTING.md
@@ -1,0 +1,161 @@
+# Contributing to `MGIS`
+
+Contributions to `MGIS` are greatly appreciated.
+
+Please take a moment to review this document in order to make the
+contribution process easy and effective for everyone involved.
+
+Following these guidelines helps to communicate that you respect the time of
+the developers managing and developing this open source project. In return,
+they should reciprocate that respect in addressing your issue or assessing
+patches and features.
+
+## Using the issue tracker
+
+The [issue
+tracker](https://github.com/thelfer/MFrontGenericInterfaceSupport/issues)
+is the preferred channel for [bug reports](#bugs), [features
+requests](#features) and [submitting pull requests](#pull-requests), but
+please respect the following restrictions:
+
+* Please **do not** use the issue tracker for personal support requests
+  (contact directly the authors
+  [tfel-contact@cea.fr](mailto:tfel-contact@cea.fr)
+
+## Bug reports
+
+A bug is a _demonstrable problem_ that is caused by the code in the repository.
+Good bug reports are extremely helpful - thank you!
+
+Guidelines for bug reports:
+
+1. **Use the GitHub issue search**: check if the issue has already been
+   reported.
+
+2. **Check if the issue has been fixed**: try to reproduce it using the
+   latest `master` or development branch in the repository.
+
+3. **Isolate the problem**: ideally create a [reduced test
+   case].
+
+A good bug report shouldn't leave others needing to chase you up for more
+information. Please try to be as detailed as possible in your report. What is
+your environment? What steps will reproduce the issue? What compiler(s) and OS
+experience the problem? What would you expect to be the outcome? All these
+details will help people to fix any potential bugs.
+
+Example:
+
+> Short and descriptive example bug report title
+>
+> A summary of the issue, versions of `TFEL/MFront` used and the
+> OS/compiler environment in which it occurs. If suitable, include the
+> steps required to reproduce the bug.
+>
+> 1. This is the first step
+> 2. This is the second step
+> 3. Further steps, etc.
+>
+> Any other information you want to share that is relevant to the issue being
+> reported. This might include the lines of code that you have identified as
+> causing the bug, and potential solutions (and your opinions on their
+> merits).
+
+## Feature requests
+
+Feature requests are welcome. But take a moment to find out whether your idea
+fits with the scope and aims of the project. It's up to *you* to make a strong
+case to convince the project's developers of the merits of this feature. Please
+provide as much detail and context as possible.
+
+## Pull requests
+
+Good pull requests - patches, improvements, new features - are a fantastic
+help. They should remain focused in scope and avoid containing unrelated
+commits.
+
+**Please ask first** before embarking on any significant pull request (e.g.
+implementing features, refactoring code, porting to a different language),
+otherwise you risk spending a lot of time working on something that the
+project's developers might not want to merge into the project.
+
+Please adhere to the coding conventions used throughout a project (indentation,
+accurate comments, etc.) and any other requirements (such as test coverage).
+
+Adhering to the following this process is the best way to get your work
+included in the project:
+
+1. [Fork](http://help.github.com/fork-a-repo/) the project, clone your fork,
+   and configure the remotes:
+
+   ```bash
+   # Clone your fork of the repo into the current directory
+   git clone https://github.com/<your-username>/MFrontGenericInterfaceSupport.git
+   # Navigate to the newly cloned directory
[…]
