+++ bindings/c/include/CMakeLists.txt
@@ -2,6 +2,7 @@ mgis_header(MGIS Config.h)
 mgis_header(MGIS Status.h)
 mgis_header(MGIS ThreadPool.h)
 mgis_header(MGIS/Behaviour Variable.h)
+mgis_header(MGIS/Behaviour Hypothesis.h)
 mgis_header(MGIS/Behaviour Behaviour.h)
 mgis_header(MGIS/Behaviour State.h)
 mgis_header(MGIS/Behaviour StateView.h)

[…]
