+++ docs/web/bibliography.bib
@@ -43,7 +43,74 @@
 	author = {Bleyer, Jérémy and Helfer, Thomas},
 	date = {2019}
 }
+@article{bourdin2000numerical,
+  title={Numerical experiments in revisited brittle fracture},
+  author={Bourdin, Blaise and Francfort, Gilles A and Marigo, Jean-Jacques},
+  journal={Journal of the Mechanics and Physics of Solids},
+  volume={48},
+  number={4},
+  pages={797--826},
+  year={2000},
+  publisher={Elsevier}
+}
 
+@Article{ambrosio1990approximation,
+  author    = {Ambrosio, Luigi and Tortorelli, Vincenzo Maria},
+  title     = {Approximation of functional depending on jumps by elliptic functional via t-convergence},
+  journal   = {Communications on Pure and Applied Mathematics},
+  year      = {1990},
+  volume    = {43},
+  number    = {8},
+  pages     = {999--1036},
+  owner     = {bleyerj},
+  publisher = {Wiley Online Library},
+  timestamp = {2017.07.04},
+}
+
+@Article{francfort1998revisiting,
+  author    = {Francfort, Gilles A and Marigo, J-J},
+  title     = {Revisiting brittle fracture as an energy minimization problem},
+  journal   = {Journal of the Mechanics and Physics of Solids},
+  year      = {1998},
+  volume    = {46},
+  number    = {8},
+  pages     = {1319--1342},
+  owner     = {bleyerj},
+  publisher = {Elsevier},
+  timestamp = {2017.07.04},
+}
+@article{amor2009regularized,
+  title={Regularized formulation of the variational brittle fracture with unilateral contact: Numerical experiments},
+  author={Amor, Hanen and Marigo, Jean-Jacques and Maurini, Corrado},
+  journal={Journal of the Mechanics and Physics of Solids},
+  volume={57},
+  number={8},
+  pages={1209--1229},
+  year={2009},
+  publisher={Elsevier}
+}
+
+@article{pham2011gradient,
+  title={Gradient damage models and their use to approximate brittle fracture},
+  author={Pham, Kim and Amor, Hanen and Marigo, Jean-Jacques and Maurini, Corrado},
+  journal={International Journal of Damage Mechanics},
+  volume={20},
+  number={4},
+  pages={618--652},
+  year={2011},
+  publisher={SAGE Publications Sage UK: London, England}
+}
+
+@article{miehe2010phase,
+  title={A phase field model for rate-independent crack propagation: Robust algorithmic implementation based on operator splits},
+  author={Miehe, Christian and Hofacker, Martina and Welschinger, Fabian},
+  journal={Computer Methods in Applied Mechanics and Engineering},
+  volume={199},
+  number={45-48},
+  pages={2765--2778},
+  year={2010},
+  publisher={Elsevier}
+}
 @article{Lester2017,
 title = {Verification of the Skorohod-Olevsky Viscous Sintering (SOVS) Model},
 author = {Lester, Brian T.},
