+++ docs/web/mgis_fenics_multiphase_model.md
@@ -13,11 +13,16 @@ $\newcommand{\bsig}{\boldsymbol{\sigma}}
 \newcommand{\div}{\operatorname{div}}
 \newcommand{\jump}[1]{[\![#1]\!]}
 \newcommand{\avg}[1]{\left\langle#1\right\rangle}$
-This demo explores the full capability of MFront's recent extension to handling generalized behaviours, namely involving different pairs of fluxes and gradients, as well as FEniCS versatility to solve generic PDEs.
+This demo explores the full capability of `MFront`'s recent extension to handling generalized behaviours, namely involving different pairs of fluxes and gradients, as well as `FEniCS` versatility to solve generic PDEs.
 
 This will be illustrated on a generalized continuum model called *multiphase model* in the context of fiber-reinforced materials [@bleyer2018multiphase].
 
 # A quick primer on the multiphase model
+
+<figure>
+<img src="supplementary_files/multiphase_kinematics.svg" width="500">
+  <figcaption>Kinematics of the multiphase model</figcaption>
+</figure>
 
 The multiphase model is a higher-grade (i.e. with enhanced kinematics) generalized continuum which represents a biphasic material (the main application being fiber-reinforced materials) by a superposition of two phases (say *matrix* and *fiber* phases), each possessing their own kinematics and in mutual interaction. Each phase kinematics is described by a displacement field $\bU^1$ and $\bU^2$ for the matrix and fiber phase respectively.
 
@@ -60,9 +65,9 @@ which will be particularized for the present demo to the following *elastic beha
 \end{align*}
 in which $\DD^{ij}$ are partial stiffness tensors and $\kappa$ can be seen as an interaction stiffness between both phases.
 
-# MFront implementation
+# `MFront` implementation
 
-The MFront implementation expands upon the detailed MFront implementation of the [stationnary heat transfer demo](mgis_fenics_nonlinear_heat_transfer.html). Again, the  `DefaultGenericBehaviour` is used here and specify that the following implementation will only handle the 2D plane strain case.
+The `MFront` implementation expands upon the detailed `MFront` implementation of the [stationnary heat transfer demo](mgis_fenics_nonlinear_heat_transfer.html). Again, the  `DefaultGenericBehaviour` is used here and specify that the following implementation will only handle the 2D plane strain case.
 ``` cpp
 @DSL DefaultGenericBehaviour;
 @Behaviour MultiphaseModel;
@@ -72,6 +77,7 @@ The MFront implementation expands upon the detailed MFront implementation of the
 @ModellingHypotheses {PlaneStrain};
 ```
 
+## Flux/Gradient dual pairs
 The three pairs of generalized flux/gradient of the multiphase model are then defined, namely $(\bsig^1,\beps^1)$, $(\bsig^2,\beps^2)$ and $(\bI,\bV)$: 
 
 ```cpp
@@ -115,6 +121,8 @@ nu2.setEntryName("FiberPoissonRatio");
 s.setEntryName("Size");
 ```
 
+## Generalized elastic behaviour
+
 It has been shown in [@bleyer2018multiphase] that, for materials made of two constituents, the partial stiffness tensors $\DD^{ij}$ can be expressed as functions of the material individual stiffness $\CC^1,\CC^2$ and the macroscopic stiffness $\CC^{hom}$ of the composite (obtained from classical homogenization theory) as follows:
 
 \begin{align*}
@@ -126,7 +134,7 @@ It has been shown in [@bleyer2018multiphase] that, for materials made of two con
 
[…]
