+++ docs/web/mgis_fenics_small_strain_elastoplasticity.md
@@ -1,15 +1,15 @@
 % Small-strain von Mises elastoplasticity
 
-This demo is the third installment in a series of two previous demos considering the implementation of von Mises elastoplasticity in FEniCS published on [Numerical tours of continuum mechanics using FEniCS](https://comet-fenics.readthedocs.io):
+This demo is the third installment in a series of two previous demos considering the implementation of von Mises elastoplasticity in ̀FEniCS` published on [Numerical tours of continuum mechanics using FEniCS](https://comet-fenics.readthedocs.io):
 
-1. [Elasto-plastic analysis of a 2D von Mises material](https://comet-fenics.readthedocs.io/en/latest/demo/2D_plasticity/vonMises_plasticity.py.html): this first installment presents a pure FEniCS implementation of von Mises plasticity which uses UFL operators for implementing the return mapping procedure. It is therefore not general since it relies heavily on the fact that the return mapping can be expressed in closed-form analytic expression in the case of von Mises plasticity with linear isotropic hardening. It nevertheless shows how to integrate the constitutive relation using `Quadrature` elements, including the use of consistent tangent operators, inside a global Newton-Raphson method.
[…]
